Confidence Intervals for Bootstrapped Model Parameters

Confidence Intervals for Bootstrapped Model Parameters

# S3 method for rendo.boots
confint(object, parm, level = 0.95, ...)

Arguments

object
a fitted model object with bootstrapped parameters. Typically from copulaCorrection

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the confidence level required.

...
ignored, for consistency with the generic function.

Details

Computes the two-sided percentile confidence intervals from the bootstrapped parameter estimates. The intervals are obtained by selecting the quantile of the bootstrapped parameter estimates corresponding to the given alpha level.

A minimum of 1/min(level, 1-level) parameters estimates are needed to derive the confidence interval. The reason for this is that there is otherwise no natural way to derive the percentiles (ie one cannot reasonably estimate the 95% quantile of only 7 values).