Multilevel Simulated Dataset - Three Levels

A dataset simulated to exemplify the use of the multilevelIV() function. It has 2767 observations, clustered into 40 level-three variables and 1347 observations at level two. The endogenous regressor is X15 with a true coefficient value of -1.

data("dataMultilevelIV")

Format

A data frame with 2767 observations clustered into 40 level-three variables and 1347 level-two variables.

y

a numeric vector representing the dependent variable.

X11

a level-one numeric vector representing a categorical exogenous variable with true parameter value equal to 3.

X12

a level-one numeric vector representing a binomial distributed exogenous variable with true parameter value equal to 9.

X13

a level-one numeric vector representing a binomial distributed exogenous variable with true parameter value equal to -2.

X14

a level-two numeric vector representing a normally distributed exogenous variable with true parameter value equal to 2.

X15

a level-two numeric vector representing a normally distributed endogenous variable, correlated with the level-two errors. It true parameter value equals to \(-1\) and it has a correlation with the level two errors equal to 0.7.

X21

a level-two numeric vector representing a binomial distributed exogenous variable with true parameter value equal to -1.5.

X22

a level-two numeric vector representing a binomial distributed exogenous variable with true parameter value equal to -4.

X23

a level-two numeric vector representing a binomial distributed exogenous variable with true parameter value equal to -3.

X24

a level-teo numeric vector representing a normally distributed exogenous variable with true parameter value equal to 6.

X31

a level-three numeric vector representing a normally distributed exogenous variable with true parameter value equal to 0.5.

X32

a level-three numeric vector representing a truncated normally distributed exogenous variable with true parameter value equal to 0.1.

X33

a level-three numeric vector representing a truncated normally distributed exogenous variable with true parameter value equal to -0.5.

SID

a numeric vector identifying each level-three observations.

CID

a numeric vector identifying each level-two observations.

Author

Raluca Gui raluca.gui@business.uzh.ch