Fitting Linear Models with one Endogenous Regressor using Latent Instrumental Variables

Fits linear models with one endogenous regressor and no additional explanatory variables using the latent instrumental variable approach presented in Ebbes, P., Wedel, M., Böckenholt, U., and Steerneman, A. G. M. (2005). This is a statistical technique to address the endogeneity problem where no external instrumental variables are needed. The important assumption of the model is that the latent variables are discrete with at least two groups with different means and the structural error is normally distributed.

latentIV(
  formula,
  data,
  start.params = c(),
  optimx.args = list(),
  verbose = TRUE
)

Arguments

formula

A symbolic description of the model to be fitted. Of class "formula".

data

A data.frame containing the data of all parts specified in the formula parameter.

start.params

A named vector containing a set of parameters to use in the first optimization iteration. The names have to correspond exactly to the names of the components specified in the formula parameter. If not provided, a linear model is fitted to derive them.

optimx.args

A named list of arguments which are passed to optimx. This allows users to tweak optimization settings to their liking.

verbose

Show details about the running of the function.

Value

An object of classes rendo.latent.IV and rendo.base is returned which is a list and contains the following components:

formula

The formula given to specify the fitted model.

terms

The terms object used for model fitting.

model

The model.frame used for model fitting.

coefficients

A named vector of all coefficients resulting from model fitting.

names.main.coefs

a vector specifying which coefficients are from the model. For internal usage.

start.params

A named vector with the initial set of parameters used to optimize the log-likelihood function.

res.optimx

The result object returned by the function optimx after optimizing the log-likelihood function.

hessian

A named, symmetric matrix giving an estimate of the Hessian at the found solution.

m.delta.diag

A diagonal matrix needed when deriving the vcov to apply the delta method on theta5 which was transformed during the LL optimization.

fitted.values

Fitted values at the found optimal solution.

residuals

The residuals at the found optimal solution.

The function summary can be used to obtain and print a summary of the results. The generic accessor functions coefficients, fitted.values, residuals, vcov, confint, logLik, AIC, BIC, case.names, and nobs are available.

Details

Let's consider the model:

Yt=β0+αPt+εt

Pt=π'Zt+νt

where \(t = 1,..,T\) indexes either time or cross-sectional units, Yt is the dependent variable, Pt is a k x 1 continuous, endogenous regressor, εt is a structural error term with mean zero and E(ε2)=σε2, \(\alpha\) and β0 are model parameters. Z;t is a l x 1 vector of instruments, and νt is a random error with mean zero and E(ν2)=σν2. The endogeneity problem arises from the correlation of \(P\) and εt through E(εν)=σεν

latentIV considers Zt' to be a latent, discrete, exogenous variable with an unknown number of groups \(m\) and \(\pi\) is a vector of group means. It is assumed that \(Z\) is independent of the error terms \(\epsilon\) and \(\nu\) and that it has at least two groups with different means. The structural and random errors are considered normally distributed with mean zero and variance-covariance matrix \(\Sigma\):

Σ=(σε2, σ02,
σ02, σν2)

The identification of the model lies in the assumption of the non-normality of Pt, the discreteness of the unobserved instruments and the existence of at least two groups with different means.

The method has been implemented such that the latent variable has two groups. Ebbes et al.(2005) show in a Monte Carlo experiment that even if the true number of the categories of the instrument is larger than two, estimates are approximately consistent. Besides, overfitting in terms of the number of groups/categories reduces the degrees of freedom and leads to efficiency loss. For a model with additional explanatory variables a Bayesian approach is needed, since in a frequentist approach identification issues appear.

Identification of the parameters relies on the distributional assumptions of the latent instruments as well as that of the endogenous regressor Pt. Specifically, the endogenous regressor should have a non-normal distribution while the unobserved instruments, \(Z\), should be discrete and have at least two groups with different means Ebbes, Wedel, and Böckenholt (2009). A continuous distribution for the instruments leads to an unidentified model, while a normal distribution of the endogenous regressor gives rise to inefficient estimates.

Additional parameters used during model fitting and printed in summary are:

pi1

The instrumental variables \(Z\) are assumed to be divided into two groups. pi1 represents the estimated group mean of the first group.

pi2

The estimated group mean of the second group of the instrumental variables \(Z\).

theta5

The probability of being in the first group of the instruments.

theta6

The variance, σε2

theta7

The covariance, σεν

theta8

The variance, σν2

References

Ebbes, P., Wedel,M., Böckenholt, U., and Steerneman, A. G. M. (2005). 'Solving and Testing for Regressor-Error (in)Dependence When no Instrumental Variables are Available: With New Evidence for the Effect of Education on Income'. Quantitative Marketing and Economics, 3:365--392.

Ebbes P., Wedel M., Böckenholt U. (2009). “Frugal IV Alternatives to Identify the Parameter for an Endogenous Regressor.” Journal of Applied Econometrics, 24(3), 446–468.

See also

Examples

# \donttest{
data("dataLatentIV")

# function call without any initial parameter values
l  <- latentIV(y ~ P, data = dataLatentIV)
#> No start parameters were given. The linear model y ~ P is fitted to derive them.
#> The start parameters c((Intercept)=2.628, P=-0.955, pi1=7.627, pi2=11.784, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.
summary(l)
#> 
#> Call:
#> latentIV(formula = y ~ P, data = dataLatentIV)
#> 
#> Coefficients:
#>             Estimate Std. Error z-score Pr(>|z|)    
#> (Intercept)  3.32351    0.37383    8.89   <2e-16 ***
#> P           -1.04580    0.04894  -21.37   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Further parameters estimated during model fitting:
#>     pi1     pi2  theta5  theta6  theta7  theta8 
#>  3.7652  8.5354  0.1905  1.1277  1.5754 13.7507 
#> (see help file for details)
#> 
#> Initial parameter values:
#> (Intercept)=2.6275 P=-0.9545 pi1=7.6273 pi2=11.7843 theta5=0.5
#> theta6=1 theta7=0.5 theta8=1
#> 
#> The value of the log-likelihood function: 10627.18 
#> AIC: -21238.35 , BIC: -21191.76 
#> KKT1: TRUE  KKT2: TRUE Optimx Convergence Code: 0 

# function call with initial parameter values given by the user
l1 <- latentIV(y ~ P, start.params = c("(Intercept)"=2.5, P=-0.5),
               data = dataLatentIV)
#> The start parameters c((Intercept)=2.5, P=-0.5, pi1=7.627, pi2=11.784, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.
summary(l1)
#> 
#> Call:
#> latentIV(formula = y ~ P, data = dataLatentIV, start.params = c(`(Intercept)` = 2.5, 
#>     P = -0.5))
#> 
#> Coefficients:
#>             Estimate Std. Error z-score Pr(>|z|)    
#> (Intercept)  3.01775    0.40997   7.361 2.47e-13 ***
#> P           -1.00585    0.05371 -18.726  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Further parameters estimated during model fitting:
#>     pi1     pi2  theta5  theta6  theta7  theta8 
#>  3.2569  8.3458  0.1417  1.0283  0.8863 14.0949 
#> (see help file for details)
#> 
#> Initial parameter values:
#> (Intercept)=2.5 P=-0.5 pi1=7.6273 pi2=11.7843 theta5=0.5
#> theta6=1 theta7=0.5 theta8=1
#> 
#> The value of the log-likelihood function: 10626.73 
#> AIC: -21237.46 , BIC: -21190.87 
#> KKT1: FALSE  KKT2: TRUE Optimx Convergence Code: 0 

# use own optimization settings (see optimx())
# set maximum number of iterations to 50'000
l2 <- latentIV(y ~ P, optimx.args = list(itnmax = 50000),
               data = dataLatentIV)
#> No start parameters were given. The linear model y ~ P is fitted to derive them.
#> The start parameters c((Intercept)=2.628, P=-0.955, pi1=7.627, pi2=11.784, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

# print detailed tracing information on progress
l3 <- latentIV(y ~ P, optimx.args = list(control = list(trace = 6)),
               data = dataLatentIV)
#> No start parameters were given. The linear model y ~ P is fitted to derive them.
#> The start parameters c((Intercept)=2.628, P=-0.955, pi1=7.627, pi2=11.784, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.
#> fn is  fn 
#> Looking for method =  Nelder-Mead 
#> Methods to be used:[1] "Nelder-Mead"
#> Function has  8  arguments
#> Analytic gradient not made available.
#> Analytic Hessian not made available.
#> Scale check -- log parameter ratio= 1.372334   log bounds ratio= NA 
#> optcfg:$fname
#> [1] "fn"
#> 
#> $npar
#> [1] 8
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 6
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] TRUE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (params, m.data.mvnorm, use.intercept, name.intercept, 
#>     name.endo.param) 
#> {
#>     if (use.intercept) {
#>         b00 <- params[name.intercept]
#>         a1 <- params[name.endo.param]
#>     }
#>     else {
#>         b00 <- 0
#>         a1 <- params[name.endo.param]
#>     }
#>     pt <- exp(params["theta5"])
#>     pt <- pt/(1 + pt)
#>     m.sigma <- matrix(c(params["theta6"], params["theta7"], params["theta7"], 
#>         params["theta8"]), byrow = TRUE, ncol = 2, nrow = 2)
#>     s2e <- m.sigma[1, 1]
#>     s2v <- m.sigma[2, 2]
#>     sev <- m.sigma[1, 2]
#>     varcov <- matrix(0, 2, 2)
#>     varcov[1, 1] <- a1 * a1 * s2v + 2 * a1 * sev + s2e
#>     varcov[2, 1] <- a1 * s2v + sev
#>     varcov[1, 2] <- varcov[2, 1]
#>     varcov[2, 2] <- s2v
#>     pi1 <- params["pi1"]
#>     mu1 <- matrix(data = c(b00 + a1 * pi1, pi1), nrow = 2, ncol = 1)
#>     pi2 <- params["pi2"]
#>     mu2 <- matrix(c(b00 + a1 * pi2, pi2), nrow = 2, ncol = 1)
#>     log.pdf1 <- dmvnorm(m.data.mvnorm, mean = mu1, sigma = varcov, 
#>         log = TRUE)
#>     log.pdf2 <- dmvnorm(m.data.mvnorm, mean = mu2, sigma = varcov, 
#>         log = TRUE)
#>     max.AB <- pmax(log(pt) + log.pdf1, log(1 - pt) + log.pdf2)
#>     logLL.lse <- sum(max.AB + log(pt * exp(log.pdf1 - max.AB) + 
#>         (1 - pt) * exp(log.pdf2 - max.AB)))
#>     return(-1 * logLL.lse)
#> }
#> <bytecode: 0x7feec79d7168>
#> <environment: namespace:REndo>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "Nelder-Mead"
#> 
#> Method:  Nelder-Mead 
#>   Nelder-Mead direct search function minimizer
#> function value for initial parameters = 25166.340666
#>   Scaled convergence tolerance is 0.000375008
#> Stepsize computed as 1.178432
#> BUILD              9 99999999999999996863366107917975552.000000 15117.921096
#> LO-REDUCTION      11 188968.490520 15117.921096
#> LO-REDUCTION      13 40767.922605 15117.921096
#> HI-REDUCTION      15 32202.843462 15117.921096
#> LO-REDUCTION      17 25286.228659 15117.921096
#> LO-REDUCTION      19 25166.340666 15117.921096
#> LO-REDUCTION      21 25025.809571 15117.921096
#> LO-REDUCTION      23 24765.258130 15117.921096
#> LO-REDUCTION      25 24749.299831 15117.921096
#> EXTENSION         27 23763.304905 13289.307985
#> LO-REDUCTION      29 22689.840233 13289.307985
#> LO-REDUCTION      31 18748.209813 13289.307985
#> LO-REDUCTION      33 17064.949547 13289.307985
#> EXTENSION         35 15949.704321 12617.403054
#> EXTENSION         37 15622.268823 12459.472104
#> LO-REDUCTION      39 15599.127400 12459.472104
#> LO-REDUCTION      41 15117.921096 12459.472104
#> REFLECTION        43 14515.890215 12173.848725
#> REFLECTION        45 14095.141973 11852.510184
#> LO-REDUCTION      47 13318.637529 11852.510184
#> HI-REDUCTION      49 13289.307985 11852.510184
#> LO-REDUCTION      51 12680.455512 11852.510184
#> LO-REDUCTION      53 12617.403054 11852.510184
#> LO-REDUCTION      55 12562.675931 11832.131827
#> LO-REDUCTION      57 12459.472104 11832.131827
#> REFLECTION        59 12186.287407 11680.397304
#> HI-REDUCTION      61 12173.848725 11680.397304
#> LO-REDUCTION      63 11966.737907 11680.178949
#> LO-REDUCTION      65 11924.068935 11680.178949
#> LO-REDUCTION      67 11900.865026 11680.178949
#> REFLECTION        69 11852.510184 11678.040110
#> REFLECTION        71 11840.556176 11634.168568
#> REFLECTION        73 11834.455510 11617.208967
#> LO-REDUCTION      75 11832.131827 11611.984653
#> LO-REDUCTION      77 11740.635088 11611.984653
#> EXTENSION         79 11687.982331 11391.769027
#> LO-REDUCTION      81 11680.397304 11391.769027
#> LO-REDUCTION      83 11680.178949 11391.769027
#> LO-REDUCTION      85 11678.040110 11391.769027
#> LO-REDUCTION      87 11660.985085 11391.769027
#> LO-REDUCTION      89 11634.168568 11391.769027
#> EXTENSION         91 11617.208967 11347.770092
#> LO-REDUCTION      93 11611.984653 11347.770092
#> EXTENSION         95 11535.227479 11169.595975
#> LO-REDUCTION      97 11515.933673 11169.595975
#> LO-REDUCTION      99 11491.243202 11169.595975
#> LO-REDUCTION     101 11477.374129 11169.595975
#> LO-REDUCTION     103 11436.389556 11169.595975
#> LO-REDUCTION     105 11401.552409 11169.595975
#> LO-REDUCTION     107 11391.769027 11169.595975
#> LO-REDUCTION     109 11347.770092 11169.595975
#> LO-REDUCTION     111 11281.108176 11160.339850
#> REFLECTION       113 11255.801268 11136.567810
#> LO-REDUCTION     115 11247.482041 11136.567810
#> LO-REDUCTION     117 11199.982234 11136.567810
#> LO-REDUCTION     119 11192.659182 11136.567810
#> HI-REDUCTION     121 11190.855303 11136.567810
#> REFLECTION       123 11172.529098 11120.441993
#> HI-REDUCTION     125 11169.595975 11120.441993
#> LO-REDUCTION     127 11160.339850 11120.441993
#> LO-REDUCTION     129 11158.662827 11120.441993
#> REFLECTION       131 11154.644746 11112.450306
#> REFLECTION       133 11153.050009 11110.959229
#> EXTENSION        135 11146.887191 11067.813800
#> LO-REDUCTION     137 11137.234238 11067.813800
#> LO-REDUCTION     139 11136.567810 11067.813800
#> LO-REDUCTION     141 11126.161822 11067.813800
#> EXTENSION        143 11122.641238 11048.881404
#> LO-REDUCTION     145 11120.441993 11048.881404
#> EXTENSION        147 11112.450306 11022.266441
#> LO-REDUCTION     149 11110.959229 11022.266441
#> EXTENSION        151 11096.257911 10987.533310
#> LO-REDUCTION     153 11090.752816 10987.533310
#> LO-REDUCTION     155 11084.237571 10987.533310
#> LO-REDUCTION     157 11074.504137 10987.533310
#> EXTENSION        159 11067.813800 10948.860089
#> EXTENSION        161 11048.881404 10880.366179
#> LO-REDUCTION     163 11045.084523 10880.366179
#> LO-REDUCTION     165 11022.266441 10880.366179
#> LO-REDUCTION     167 11006.951360 10880.366179
#> LO-REDUCTION     169 11005.515754 10880.366179
#> LO-REDUCTION     171 10990.570205 10880.366179
#> EXTENSION        173 10987.533310 10848.335611
#> LO-REDUCTION     175 10948.860089 10848.335611
#> EXTENSION        177 10944.462527 10807.847832
#> LO-REDUCTION     179 10912.389163 10807.847832
#> EXTENSION        181 10907.118651 10778.654925
#> LO-REDUCTION     183 10906.411147 10778.654925
#> EXTENSION        185 10902.247761 10773.831766
#> REFLECTION       187 10880.366179 10751.324420
#> LO-REDUCTION     189 10874.504867 10751.324420
#> REFLECTION       191 10848.335611 10728.000065
#> LO-REDUCTION     193 10815.565359 10728.000065
#> LO-REDUCTION     195 10807.847832 10728.000065
#> EXTENSION        197 10782.821843 10689.847737
#> LO-REDUCTION     199 10778.654925 10689.847737
#> LO-REDUCTION     201 10773.831766 10689.847737
#> LO-REDUCTION     203 10763.291160 10689.847737
#> LO-REDUCTION     205 10751.324420 10689.847737
#> LO-REDUCTION     207 10746.415029 10689.847737
#> REFLECTION       209 10737.260086 10687.950253
#> REFLECTION       211 10728.000065 10683.155727
#> LO-REDUCTION     213 10720.881027 10683.155727
#> REFLECTION       215 10715.153721 10678.176674
#> REFLECTION       217 10711.924187 10677.771425
#> LO-REDUCTION     219 10702.816314 10677.771425
#> LO-REDUCTION     221 10700.833332 10677.771425
#> LO-REDUCTION     223 10693.943698 10677.771425
#> REFLECTION       225 10689.847737 10667.695390
#> LO-REDUCTION     227 10687.950253 10667.695390
#> LO-REDUCTION     229 10683.996776 10667.695390
#> EXTENSION        231 10683.155727 10660.286808
#> HI-REDUCTION     233 10678.568899 10660.286808
#> LO-REDUCTION     235 10678.176674 10660.286808
#> LO-REDUCTION     237 10677.778519 10660.286808
#> REFLECTION       239 10677.771425 10660.116619
#> LO-REDUCTION     241 10671.518869 10660.116619
#> EXTENSION        243 10670.140705 10654.024464
#> LO-REDUCTION     245 10669.955114 10654.024464
#> REFLECTION       247 10667.695390 10653.410447
#> LO-REDUCTION     249 10664.803084 10653.410447
#> EXTENSION        251 10663.099452 10644.447977
#> LO-REDUCTION     253 10662.231290 10644.447977
#> LO-REDUCTION     255 10660.286808 10644.447977
#> LO-REDUCTION     257 10660.116619 10644.447977
#> EXTENSION        259 10656.725550 10637.194156
#> LO-REDUCTION     261 10654.024464 10637.194156
#> LO-REDUCTION     263 10653.890565 10637.194156
#> LO-REDUCTION     265 10653.410447 10637.194156
#> LO-REDUCTION     267 10651.664821 10637.194156
#> LO-REDUCTION     269 10648.441063 10637.194156
#> LO-REDUCTION     271 10648.226824 10637.194156
#> LO-REDUCTION     273 10647.511960 10637.194156
#> REFLECTION       275 10644.447977 10636.991308
#> LO-REDUCTION     277 10644.315328 10636.991308
#> REFLECTION       279 10644.142298 10636.502245
#> REFLECTION       281 10641.527593 10634.534306
#> LO-REDUCTION     283 10639.177202 10634.534306
#> LO-REDUCTION     285 10639.083902 10634.534306
#> LO-REDUCTION     287 10639.017011 10634.534306
#> REFLECTION       289 10637.897839 10633.909038
#> LO-REDUCTION     291 10637.194156 10633.909038
#> LO-REDUCTION     293 10636.991308 10633.909038
#> LO-REDUCTION     295 10636.502245 10633.909038
#> LO-REDUCTION     297 10636.384966 10633.909038
#> LO-REDUCTION     299 10636.003406 10633.909038
#> LO-REDUCTION     301 10635.102824 10633.863056
#> LO-REDUCTION     303 10635.071382 10633.863056
#> REFLECTION       305 10634.734367 10633.554391
#> LO-REDUCTION     307 10634.730202 10633.554391
#> EXTENSION        309 10634.534306 10632.463033
#> LO-REDUCTION     311 10634.306774 10632.463033
#> LO-REDUCTION     313 10634.036901 10632.463033
#> LO-REDUCTION     315 10633.940853 10632.463033
#> LO-REDUCTION     317 10633.909038 10632.463033
#> EXTENSION        319 10633.863056 10631.676687
#> LO-REDUCTION     321 10633.581683 10631.676687
#> LO-REDUCTION     323 10633.554391 10631.676687
#> LO-REDUCTION     325 10633.372894 10631.676687
#> LO-REDUCTION     327 10633.096476 10631.676687
#> LO-REDUCTION     329 10633.095047 10631.676687
#> LO-REDUCTION     331 10633.008090 10631.676687
#> EXTENSION        333 10632.705129 10630.936606
#> LO-REDUCTION     335 10632.553326 10630.936606
#> LO-REDUCTION     337 10632.551100 10630.936606
#> EXTENSION        339 10632.463033 10630.323066
#> LO-REDUCTION     341 10632.318709 10630.323066
#> LO-REDUCTION     343 10631.845863 10630.323066
#> LO-REDUCTION     345 10631.810527 10630.323066
#> LO-REDUCTION     347 10631.676687 10630.323066
#> EXTENSION        349 10631.227544 10629.764518
#> LO-REDUCTION     351 10631.191103 10629.764518
#> LO-REDUCTION     353 10630.936606 10629.764518
#> LO-REDUCTION     355 10630.628129 10629.764518
#> EXTENSION        357 10630.599965 10629.261303
#> LO-REDUCTION     359 10630.541696 10629.261303
#> LO-REDUCTION     361 10630.378393 10629.261303
#> LO-REDUCTION     363 10630.323066 10629.261303
#> LO-REDUCTION     365 10630.194936 10629.261303
#> EXTENSION        367 10630.094019 10628.940512
#> LO-REDUCTION     369 10629.986547 10628.940512
#> LO-REDUCTION     371 10629.793089 10628.940512
#> LO-REDUCTION     373 10629.764518 10628.940512
#> EXTENSION        375 10629.748821 10628.374920
#> LO-REDUCTION     377 10629.670716 10628.374920
#> LO-REDUCTION     379 10629.360664 10628.374920
#> LO-REDUCTION     381 10629.327032 10628.374920
#> LO-REDUCTION     383 10629.261303 10628.374920
#> LO-REDUCTION     385 10629.219663 10628.374920
#> EXTENSION        387 10629.016933 10628.077242
#> LO-REDUCTION     389 10628.940512 10628.077242
#> LO-REDUCTION     391 10628.756769 10628.077242
#> EXTENSION        393 10628.729848 10627.808834
#> LO-REDUCTION     395 10628.586106 10627.808834
#> LO-REDUCTION     397 10628.554895 10627.808834
#> LO-REDUCTION     399 10628.446521 10627.808834
#> REFLECTION       401 10628.374920 10627.802181
#> LO-REDUCTION     403 10628.120818 10627.802181
#> HI-REDUCTION     405 10628.082864 10627.802181
#> LO-REDUCTION     407 10628.077242 10627.802181
#> REFLECTION       409 10627.968980 10627.714056
#> LO-REDUCTION     411 10627.964578 10627.714056
#> LO-REDUCTION     413 10627.916160 10627.714056
#> REFLECTION       415 10627.915811 10627.695307
#> LO-REDUCTION     417 10627.835597 10627.695307
#> LO-REDUCTION     419 10627.824625 10627.695307
#> LO-REDUCTION     421 10627.808834 10627.695307
#> LO-REDUCTION     423 10627.802181 10627.695307
#> LO-REDUCTION     425 10627.785784 10627.695307
#> REFLECTION       427 10627.766184 10627.690481
#> LO-REDUCTION     429 10627.752855 10627.690481
#> REFLECTION       431 10627.746218 10627.689459
#> EXTENSION        433 10627.730993 10627.668555
#> LO-REDUCTION     435 10627.729669 10627.668555
#> EXTENSION        437 10627.722652 10627.606860
#> LO-REDUCTION     439 10627.714056 10627.606860
#> LO-REDUCTION     441 10627.707588 10627.606860
#> LO-REDUCTION     443 10627.695307 10627.606860
#> LO-REDUCTION     445 10627.690481 10627.606860
#> LO-REDUCTION     447 10627.689459 10627.606860
#> LO-REDUCTION     449 10627.684827 10627.606860
#> LO-REDUCTION     451 10627.668555 10627.606860
#> LO-REDUCTION     453 10627.650355 10627.606860
#> LO-REDUCTION     455 10627.649087 10627.606860
#> REFLECTION       457 10627.648278 10627.598382
#> EXTENSION        459 10627.639806 10627.546402
#> LO-REDUCTION     461 10627.633180 10627.546402
#> LO-REDUCTION     463 10627.622030 10627.546402
#> EXTENSION        465 10627.615643 10627.517784
#> LO-REDUCTION     467 10627.609777 10627.517784
#> LO-REDUCTION     469 10627.609287 10627.517784
#> LO-REDUCTION     471 10627.606860 10627.517784
#> EXTENSION        473 10627.598382 10627.505271
#> EXTENSION        475 10627.575663 10627.445320
#> LO-REDUCTION     477 10627.561964 10627.445320
#> LO-REDUCTION     479 10627.553455 10627.445320
#> LO-REDUCTION     481 10627.546402 10627.445320
#> REFLECTION       483 10627.545679 10627.439659
#> LO-REDUCTION     485 10627.518663 10627.439659
#> EXTENSION        487 10627.517784 10627.410904
#> LO-REDUCTION     489 10627.505271 10627.410904
#> LO-REDUCTION     491 10627.460400 10627.410904
#> LO-REDUCTION     493 10627.457991 10627.410904
#> LO-REDUCTION     495 10627.455565 10627.410904
#> LO-REDUCTION     497 10627.446902 10627.410904
#> LO-REDUCTION     499 10627.445320 10627.410904
#> REFLECTION       501 10627.439659 10627.410696
#> LO-REDUCTION     503 10627.428972 10627.410696
#> REFLECTION       505 10627.423307 10627.404253
#> REFLECTION       507 10627.419460 10627.402298
#> LO-REDUCTION     509 10627.418025 10627.402298
#> HI-REDUCTION     511 10627.413552 10627.402298
#> REFLECTION       513 10627.412535 10627.401685
#> LO-REDUCTION     515 10627.412192 10627.401125
#> REFLECTION       517 10627.410904 10627.399249
#> REFLECTION       519 10627.410696 10627.399140
#> LO-REDUCTION     521 10627.406126 10627.399140
#> HI-REDUCTION     523 10627.404253 10627.399140
#> LO-REDUCTION     525 10627.403722 10627.398511
#> REFLECTION       527 10627.402298 10627.395709
#> LO-REDUCTION     529 10627.401685 10627.395709
#> LO-REDUCTION     531 10627.401125 10627.395709
#> EXTENSION        533 10627.400552 10627.392989
#> LO-REDUCTION     535 10627.399609 10627.392989
#> LO-REDUCTION     537 10627.399249 10627.392989
#> HI-REDUCTION     539 10627.399140 10627.392989
#> LO-REDUCTION     541 10627.398511 10627.392989
#> REFLECTION       543 10627.396414 10627.391983
#> LO-REDUCTION     545 10627.395934 10627.391983
#> LO-REDUCTION     547 10627.395709 10627.391983
#> REFLECTION       549 10627.395233 10627.391410
#> REFLECTION       551 10627.393935 10627.390689
#> LO-REDUCTION     553 10627.393316 10627.390689
#> HI-REDUCTION     555 10627.393288 10627.390689
#> LO-REDUCTION     557 10627.393284 10627.390689
#> EXTENSION        559 10627.392989 10627.388266
#> LO-REDUCTION     561 10627.392976 10627.388266
#> LO-REDUCTION     563 10627.391983 10627.388266
#> LO-REDUCTION     565 10627.391640 10627.388266
#> LO-REDUCTION     567 10627.391410 10627.388266
#> LO-REDUCTION     569 10627.390901 10627.388266
#> EXTENSION        571 10627.390818 10627.385684
#> LO-REDUCTION     573 10627.390689 10627.385684
#> LO-REDUCTION     575 10627.390471 10627.385684
#> LO-REDUCTION     577 10627.389393 10627.385684
#> EXTENSION        579 10627.389168 10627.382827
#> LO-REDUCTION     581 10627.388941 10627.382827
#> LO-REDUCTION     583 10627.388717 10627.382827
#> LO-REDUCTION     585 10627.388266 10627.382827
#> REFLECTION       587 10627.386756 10627.382311
#> EXTENSION        589 10627.386062 10627.380832
#> LO-REDUCTION     591 10627.385881 10627.380832
#> EXTENSION        593 10627.385684 10627.376148
#> LO-REDUCTION     595 10627.384892 10627.376148
#> LO-REDUCTION     597 10627.383391 10627.376148
#> LO-REDUCTION     599 10627.383251 10627.376148
#> LO-REDUCTION     601 10627.382827 10627.376148
#> LO-REDUCTION     603 10627.382311 10627.376148
#> EXTENSION        605 10627.381803 10627.372125
#> LO-REDUCTION     607 10627.380832 10627.372125
#> EXTENSION        609 10627.379810 10627.368896
#> LO-REDUCTION     611 10627.377858 10627.368896
#> LO-REDUCTION     613 10627.377814 10627.368896
#> LO-REDUCTION     615 10627.377429 10627.368896
#> EXTENSION        617 10627.376836 10627.367220
#> EXTENSION        619 10627.376148 10627.359907
#> LO-REDUCTION     621 10627.373550 10627.359907
#> LO-REDUCTION     623 10627.372125 10627.359907
#> LO-REDUCTION     625 10627.371566 10627.359907
#> LO-REDUCTION     627 10627.370096 10627.359907
#> LO-REDUCTION     629 10627.370093 10627.359907
#> LO-REDUCTION     631 10627.368896 10627.359907
#> LO-REDUCTION     633 10627.367220 10627.359907
#> REFLECTION       635 10627.365603 10627.358847
#> REFLECTION       637 10627.363033 10627.358078
#> LO-REDUCTION     639 10627.362756 10627.358078
#> REFLECTION       641 10627.362158 10627.357725
#> EXTENSION        643 10627.361979 10627.354815
#> LO-REDUCTION     645 10627.361206 10627.354815
#> EXTENSION        647 10627.360150 10627.351066
#> LO-REDUCTION     649 10627.359907 10627.351066
#> LO-REDUCTION     651 10627.358847 10627.351066
#> LO-REDUCTION     653 10627.358437 10627.351066
#> LO-REDUCTION     655 10627.358078 10627.351066
#> EXTENSION        657 10627.357725 10627.346354
#> LO-REDUCTION     659 10627.355918 10627.346354
#> LO-REDUCTION     661 10627.355329 10627.346354
#> EXTENSION        663 10627.354815 10627.344191
#> LO-REDUCTION     665 10627.354583 10627.344191
#> EXTENSION        667 10627.353059 10627.341950
#> EXTENSION        669 10627.351480 10627.339530
#> LO-REDUCTION     671 10627.351066 10627.339530
#> LO-REDUCTION     673 10627.349584 10627.339530
#> REFLECTION       675 10627.348064 10627.338448
#> REFLECTION       677 10627.347092 10627.337467
#> LO-REDUCTION     679 10627.346354 10627.337467
#> REFLECTION       681 10627.344191 10627.337053
#> LO-REDUCTION     683 10627.341950 10627.337053
#> LO-REDUCTION     685 10627.340872 10627.337053
#> EXTENSION        687 10627.339663 10627.335305
#> LO-REDUCTION     689 10627.339530 10627.335305
#> LO-REDUCTION     691 10627.339514 10627.335305
#> LO-REDUCTION     693 10627.338448 10627.335305
#> LO-REDUCTION     695 10627.337467 10627.335305
#> LO-REDUCTION     697 10627.337363 10627.335305
#> HI-REDUCTION     699 10627.337178 10627.335305
#> LO-REDUCTION     701 10627.337173 10627.335305
#> LO-REDUCTION     703 10627.337053 10627.335305
#> LO-REDUCTION     705 10627.336694 10627.335305
#> LO-REDUCTION     707 10627.336463 10627.335246
#> REFLECTION       709 10627.336117 10627.334766
#> LO-REDUCTION     711 10627.335746 10627.334766
#> LO-REDUCTION     713 10627.335718 10627.334766
#> LO-REDUCTION     715 10627.335416 10627.334766
#> EXTENSION        717 10627.335399 10627.334454
#> LO-REDUCTION     719 10627.335376 10627.334454
#> HI-REDUCTION     721 10627.335354 10627.334454
#> EXTENSION        723 10627.335305 10627.334011
#> LO-REDUCTION     725 10627.335246 10627.334011
#> REFLECTION       727 10627.334980 10627.334010
#> EXTENSION        729 10627.334909 10627.333815
#> EXTENSION        731 10627.334812 10627.333557
#> REFLECTION       733 10627.334766 10627.333464
#> REFLECTION       735 10627.334695 10627.333453
#> LO-REDUCTION     737 10627.334454 10627.333453
#> LO-REDUCTION     739 10627.334357 10627.333401
#> EXTENSION        741 10627.334011 10627.332599
#> LO-REDUCTION     743 10627.334010 10627.332599
#> LO-REDUCTION     745 10627.333815 10627.332599
#> EXTENSION        747 10627.333580 10627.332370
#> LO-REDUCTION     749 10627.333557 10627.332370
#> LO-REDUCTION     751 10627.333464 10627.332370
#> LO-REDUCTION     753 10627.333453 10627.332370
#> LO-REDUCTION     755 10627.333401 10627.332370
#> REFLECTION       757 10627.333365 10627.332208
#> REFLECTION       759 10627.333009 10627.331942
#> EXTENSION        761 10627.332662 10627.331485
#> LO-REDUCTION     763 10627.332599 10627.331485
#> EXTENSION        765 10627.332535 10627.330805
#> LO-REDUCTION     767 10627.332457 10627.330805
#> LO-REDUCTION     769 10627.332412 10627.330805
#> LO-REDUCTION     771 10627.332370 10627.330805
#> EXTENSION        773 10627.332208 10627.330512
#> EXTENSION        775 10627.331942 10627.330315
#> REFLECTION       777 10627.331697 10627.330146
#> EXTENSION        779 10627.331503 10627.328625
#> LO-REDUCTION     781 10627.331485 10627.328625
#> EXTENSION        783 10627.331034 10627.326730
#> LO-REDUCTION     785 10627.330839 10627.326730
#> LO-REDUCTION     787 10627.330805 10627.326730
#> LO-REDUCTION     789 10627.330512 10627.326730
#> LO-REDUCTION     791 10627.330315 10627.326730
#> LO-REDUCTION     793 10627.330146 10627.326730
#> EXTENSION        795 10627.329446 10627.324630
#> LO-REDUCTION     797 10627.328625 10627.324630
#> LO-REDUCTION     799 10627.328236 10627.324630
#> LO-REDUCTION     801 10627.328118 10627.324630
#> EXTENSION        803 10627.327980 10627.322469
#> LO-REDUCTION     805 10627.327401 10627.322469
#> LO-REDUCTION     807 10627.326810 10627.322469
#> LO-REDUCTION     809 10627.326730 10627.322469
#> EXTENSION        811 10627.325512 10627.321607
#> EXTENSION        813 10627.325248 10627.319269
#> LO-REDUCTION     815 10627.325220 10627.319269
#> LO-REDUCTION     817 10627.324630 10627.319269
#> EXTENSION        819 10627.322822 10627.317764
#> LO-REDUCTION     821 10627.322747 10627.317764
#> LO-REDUCTION     823 10627.322541 10627.317764
#> LO-REDUCTION     825 10627.322469 10627.317764
#> LO-REDUCTION     827 10627.321607 10627.317764
#> REFLECTION       829 10627.321336 10627.317229
#> EXTENSION        831 10627.319403 10627.315931
#> LO-REDUCTION     833 10627.319269 10627.315931
#> HI-REDUCTION     835 10627.319217 10627.315931
#> EXTENSION        837 10627.318902 10627.313930
#> LO-REDUCTION     839 10627.318264 10627.313930
#> LO-REDUCTION     841 10627.317916 10627.313930
#> LO-REDUCTION     843 10627.317764 10627.313930
#> LO-REDUCTION     845 10627.317466 10627.313930
#> LO-REDUCTION     847 10627.317229 10627.313930
#> LO-REDUCTION     849 10627.316993 10627.313930
#> EXTENSION        851 10627.315931 10627.312988
#> LO-REDUCTION     853 10627.315200 10627.312988
#> EXTENSION        855 10627.315066 10627.311820
#> LO-REDUCTION     857 10627.315001 10627.311820
#> LO-REDUCTION     859 10627.314979 10627.311820
#> LO-REDUCTION     861 10627.314928 10627.311820
#> REFLECTION       863 10627.314415 10627.311253
#> EXTENSION        865 10627.313930 10627.308654
#> LO-REDUCTION     867 10627.313748 10627.308654
#> LO-REDUCTION     869 10627.312988 10627.308654
#> LO-REDUCTION     871 10627.312765 10627.308654
#> EXTENSION        873 10627.312247 10627.306988
#> LO-REDUCTION     875 10627.311849 10627.306988
#> LO-REDUCTION     877 10627.311820 10627.306988
#> LO-REDUCTION     879 10627.311253 10627.306988
#> EXTENSION        881 10627.311163 10627.306650
#> EXTENSION        883 10627.310029 10627.303816
#> LO-REDUCTION     885 10627.309940 10627.303816
#> LO-REDUCTION     887 10627.308654 10627.303816
#> LO-REDUCTION     889 10627.307925 10627.303816
#> LO-REDUCTION     891 10627.307508 10627.303816
#> LO-REDUCTION     893 10627.307213 10627.303816
#> LO-REDUCTION     895 10627.306988 10627.303816
#> LO-REDUCTION     897 10627.306650 10627.303816
#> EXTENSION        899 10627.306398 10627.302838
#> EXTENSION        901 10627.305260 10627.302399
#> EXTENSION        903 10627.304944 10627.301617
#> EXTENSION        905 10627.304918 10627.300176
#> LO-REDUCTION     907 10627.304681 10627.300176
#> LO-REDUCTION     909 10627.304629 10627.300176
#> LO-REDUCTION     911 10627.304257 10627.300176
#> LO-REDUCTION     913 10627.303816 10627.300176
#> LO-REDUCTION     915 10627.302838 10627.300176
#> LO-REDUCTION     917 10627.302399 10627.300176
#> LO-REDUCTION     919 10627.301901 10627.300176
#> REFLECTION       921 10627.301617 10627.300045
#> REFLECTION       923 10627.300791 10627.299395
#> EXTENSION        925 10627.300638 10627.299154
#> EXTENSION        927 10627.300468 10627.298714
#> EXTENSION        929 10627.300411 10627.298328
#> LO-REDUCTION     931 10627.300385 10627.298328
#> EXTENSION        933 10627.300323 10627.297052
#> LO-REDUCTION     935 10627.300176 10627.297052
#> EXTENSION        937 10627.300045 10627.294937
#> LO-REDUCTION     939 10627.299395 10627.294937
#> LO-REDUCTION     941 10627.299154 10627.294937
#> LO-REDUCTION     943 10627.298714 10627.294937
#> EXTENSION        945 10627.298705 10627.293423
#> EXTENSION        947 10627.298328 10627.291961
#> EXTENSION        949 10627.297271 10627.290610
#> EXTENSION        951 10627.297052 10627.286569
#> LO-REDUCTION     953 10627.295764 10627.286569
#> LO-REDUCTION     955 10627.295544 10627.286569
#> LO-REDUCTION     957 10627.295383 10627.286569
#> EXTENSION        959 10627.294937 10627.284935
#> REFLECTION       961 10627.293423 10627.283913
#> EXTENSION        963 10627.291961 10627.275925
#> LO-REDUCTION     965 10627.290610 10627.275925
#> LO-REDUCTION     967 10627.288413 10627.275925
#> LO-REDUCTION     969 10627.287854 10627.275925
#> EXTENSION        971 10627.286822 10627.270579
#> LO-REDUCTION     973 10627.286569 10627.270579
#> LO-REDUCTION     975 10627.284935 10627.270579
#> EXTENSION        977 10627.283913 10627.266499
#> EXTENSION        979 10627.280871 10627.261222
#> EXTENSION        981 10627.278905 10627.254687
#> LO-REDUCTION     983 10627.276641 10627.254687
#> LO-REDUCTION     985 10627.276111 10627.254687
#> LO-REDUCTION     987 10627.275925 10627.254687
#> EXTENSION        989 10627.274312 10627.248777
#> LO-REDUCTION     991 10627.270579 10627.248777
#> EXTENSION        993 10627.266499 10627.239120
#> LO-REDUCTION     995 10627.262812 10627.239120
#> EXTENSION        997 10627.261222 10627.232039
#> LO-REDUCTION     999 10627.255470 10627.232039
#> LO-REDUCTION    1001 10627.255433 10627.232039
#> LO-REDUCTION    1003 10627.254687 10627.232039
#> LO-REDUCTION    1005 10627.252724 10627.232039
#> LO-REDUCTION    1007 10627.249429 10627.232039
#> EXTENSION       1009 10627.248777 10627.226174
#> EXTENSION       1011 10627.240173 10627.222003
#> EXTENSION       1013 10627.239155 10627.217054
#> LO-REDUCTION    1015 10627.239120 10627.217054
#> LO-REDUCTION    1017 10627.238006 10627.217054
#> EXTENSION       1019 10627.235574 10627.210352
#> LO-REDUCTION    1021 10627.234465 10627.210352
#> LO-REDUCTION    1023 10627.232039 10627.210352
#> LO-REDUCTION    1025 10627.226174 10627.210352
#> LO-REDUCTION    1027 10627.224201 10627.210352
#> LO-REDUCTION    1029 10627.222003 10627.210352
#> LO-REDUCTION    1031 10627.220416 10627.210352
#> LO-REDUCTION    1033 10627.218157 10627.210352
#> EXTENSION       1035 10627.217054 10627.207828
#> LO-REDUCTION    1037 10627.214506 10627.207828
#> EXTENSION       1039 10627.213703 10627.203439
#> LO-REDUCTION    1041 10627.212896 10627.203439
#> LO-REDUCTION    1043 10627.212268 10627.203439
#> LO-REDUCTION    1045 10627.211161 10627.203439
#> LO-REDUCTION    1047 10627.210580 10627.203439
#> LO-REDUCTION    1049 10627.210352 10627.203439
#> REFLECTION      1051 10627.209093 10627.203206
#> LO-REDUCTION    1053 10627.207828 10627.203206
#> LO-REDUCTION    1055 10627.207380 10627.203206
#> EXTENSION       1057 10627.206098 10627.200958
#> LO-REDUCTION    1059 10627.205913 10627.200958
#> EXTENSION       1061 10627.205291 10627.197698
#> LO-REDUCTION    1063 10627.204766 10627.197698
#> LO-REDUCTION    1065 10627.204707 10627.197698
#> LO-REDUCTION    1067 10627.203603 10627.197698
#> LO-REDUCTION    1069 10627.203439 10627.197698
#> LO-REDUCTION    1071 10627.203206 10627.197698
#> EXTENSION       1073 10627.201828 10627.197264
#> EXTENSION       1075 10627.201743 10627.194831
#> LO-REDUCTION    1077 10627.201033 10627.194831
#> LO-REDUCTION    1079 10627.200958 10627.194831
#> REFLECTION      1081 10627.199721 10627.194540
#> EXTENSION       1083 10627.199665 10627.189501
#> LO-REDUCTION    1085 10627.198372 10627.189501
#> LO-REDUCTION    1087 10627.197698 10627.189501
#> LO-REDUCTION    1089 10627.197264 10627.189501
#> LO-REDUCTION    1091 10627.196167 10627.189501
#> LO-REDUCTION    1093 10627.195161 10627.189501
#> LO-REDUCTION    1095 10627.194831 10627.189501
#> EXTENSION       1097 10627.194540 10627.187810
#> EXTENSION       1099 10627.193176 10627.185764
#> LO-REDUCTION    1101 10627.192385 10627.185764
#> EXTENSION       1103 10627.191780 10627.183839
#> LO-REDUCTION    1105 10627.191764 10627.183839
#> EXTENSION       1107 10627.191712 10627.180294
#> LO-REDUCTION    1109 10627.189708 10627.180294
#> LO-REDUCTION    1111 10627.189501 10627.180294
#> LO-REDUCTION    1113 10627.187810 10627.180294
#> LO-REDUCTION    1115 10627.187378 10627.180294
#> REFLECTION      1117 10627.185806 10627.179463
#> LO-REDUCTION    1119 10627.185764 10627.179463
#> LO-REDUCTION    1121 10627.183839 10627.179463
#> LO-REDUCTION    1123 10627.183223 10627.179463
#> REFLECTION      1125 10627.182806 10627.179123
#> EXTENSION       1127 10627.180855 10627.176981
#> LO-REDUCTION    1129 10627.180691 10627.176981
#> LO-REDUCTION    1131 10627.180673 10627.176981
#> LO-REDUCTION    1133 10627.180554 10627.176981
#> LO-REDUCTION    1135 10627.180484 10627.176981
#> REFLECTION      1137 10627.180294 10627.176882
#> LO-REDUCTION    1139 10627.179463 10627.176882
#> LO-REDUCTION    1141 10627.179123 10627.176882
#> REFLECTION      1143 10627.178788 10627.176405
#> LO-REDUCTION    1145 10627.178294 10627.176405
#> REFLECTION      1147 10627.178071 10627.176354
#> LO-REDUCTION    1149 10627.177274 10627.176354
#> LO-REDUCTION    1151 10627.177200 10627.176354
#> EXTENSION       1153 10627.177182 10627.176124
#> LO-REDUCTION    1155 10627.176981 10627.176124
#> REFLECTION      1157 10627.176882 10627.176087
#> LO-REDUCTION    1159 10627.176791 10627.176087
#> LO-REDUCTION    1161 10627.176645 10627.176087
#> LO-REDUCTION    1163 10627.176632 10627.176087
#> REFLECTION      1165 10627.176499 10627.176051
#> LO-REDUCTION    1167 10627.176405 10627.176029
#> Exiting from Nelder Mead minimizer
#>     1169 function evaluations used
#> Post processing for method  Nelder-Mead 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  Nelder-Mead 
#> Compute gradient approximation at finish of  Nelder-Mead 
#> Save results from method  Nelder-Mead 
#> $par
#> (Intercept)           P         pi1         pi2      theta5      theta6 
#>    3.323506   -1.045803    3.765180    8.535425   -1.446987    1.127709 
#>      theta7      theta8 
#>    1.575367   13.750670 
#> 
#> $value
#> [1] 10627.18
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>     1169 
#> 
#> $gevals
#> gradient 
#>       NA 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] TRUE
#> 
#> $xtimes
#> user.self 
#>     0.974 
#> 
#> Assemble the answers

# use method L-BFGS-B instead of Nelder-Mead and print report every 50 iterations
l4 <- latentIV(y ~ P, optimx.args = list(method = "L-BFGS-B", control=list(trace = 2, REPORT=50)),
               data = dataLatentIV)
#> No start parameters were given. The linear model y ~ P is fitted to derive them.
#> The start parameters c((Intercept)=2.628, P=-0.955, pi1=7.627, pi2=11.784, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.
#> fn is  fn 
#> Looking for method =  L-BFGS-B 
#> Methods to be used:[1] "L-BFGS-B"
#> Function has  8  arguments
#> Analytic gradient not made available.
#> Analytic Hessian not made available.
#> Scale check -- log parameter ratio= 1.372334   log bounds ratio= NA 
#> optcfg:$fname
#> [1] "fn"
#> 
#> $npar
#> [1] 8
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 2
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] TRUE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$REPORT
#> [1] 50
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (params, m.data.mvnorm, use.intercept, name.intercept, 
#>     name.endo.param) 
#> {
#>     if (use.intercept) {
#>         b00 <- params[name.intercept]
#>         a1 <- params[name.endo.param]
#>     }
#>     else {
#>         b00 <- 0
#>         a1 <- params[name.endo.param]
#>     }
#>     pt <- exp(params["theta5"])
#>     pt <- pt/(1 + pt)
#>     m.sigma <- matrix(c(params["theta6"], params["theta7"], params["theta7"], 
#>         params["theta8"]), byrow = TRUE, ncol = 2, nrow = 2)
#>     s2e <- m.sigma[1, 1]
#>     s2v <- m.sigma[2, 2]
#>     sev <- m.sigma[1, 2]
#>     varcov <- matrix(0, 2, 2)
#>     varcov[1, 1] <- a1 * a1 * s2v + 2 * a1 * sev + s2e
#>     varcov[2, 1] <- a1 * s2v + sev
#>     varcov[1, 2] <- varcov[2, 1]
#>     varcov[2, 2] <- s2v
#>     pi1 <- params["pi1"]
#>     mu1 <- matrix(data = c(b00 + a1 * pi1, pi1), nrow = 2, ncol = 1)
#>     pi2 <- params["pi2"]
#>     mu2 <- matrix(c(b00 + a1 * pi2, pi2), nrow = 2, ncol = 1)
#>     log.pdf1 <- dmvnorm(m.data.mvnorm, mean = mu1, sigma = varcov, 
#>         log = TRUE)
#>     log.pdf2 <- dmvnorm(m.data.mvnorm, mean = mu2, sigma = varcov, 
#>         log = TRUE)
#>     max.AB <- pmax(log(pt) + log.pdf1, log(1 - pt) + log.pdf2)
#>     logLL.lse <- sum(max.AB + log(pt * exp(log.pdf1 - max.AB) + 
#>         (1 - pt) * exp(log.pdf2 - max.AB)))
#>     return(-1 * logLL.lse)
#> }
#> <bytecode: 0x7feec79d7168>
#> <environment: namespace:REndo>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "L-BFGS-B"
#> 
#> Method:  L-BFGS-B 
#> N = 8, M = 5 machine precision = 2.22045e-16
#> This problem is unconstrained.
#> 
#> iterations 98
#> function evaluations 124
#> segments explored during Cauchy searches 1
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 0
#> norm of the final projected gradient 0.238076
#> final function value 10626.7
#> 
#> final  value 10626.728237 
#> converged
#> Post processing for method  L-BFGS-B 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  L-BFGS-B 
#> Compute gradient approximation at finish of  L-BFGS-B 
#> Save results from method  L-BFGS-B 
#> $par
#> (Intercept)           P         pi1         pi2      theta5      theta6 
#>   2.9056709  -0.9909931   2.9140188   8.2497343  -2.0272287   1.0052473 
#>      theta7      theta8 
#>   0.6307542  14.3456995 
#> 
#> $value
#> [1] 10626.73
#> 
#> $message
#> [1] "CONVERGENCE: REL_REDUCTION_OF_F <= FACTR*EPSMCH"
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>      124 
#> 
#> $gevals
#> gradient 
#>      124 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] FALSE
#> 
#> $xtimes
#> user.self 
#>     1.767 
#> 
#> Assemble the answers

# read out all coefficients, incl auxiliary coefs
lat.all.coefs <- coef(l4)
# same as above
lat.all.coefs <- coef(l4, complete = TRUE)
# only main model coefs
lat.main.coefs <- coef(l4, complete = FALSE)
# }