Predict method for fitted Regression Models with Internal Instrumental Variables

Predicted values based on model objects fitted using the instrumental variables regression fitted with IVs generated from the data.

# S3 method for rendo.ivreg
predict(object, newdata, ...)

Arguments

object
Object of class inheriting from "rendo.ivreg"

newdata
An optional data frame without any instrumental variables in which to look for variables with which to predict. If omitted, the fitted values are returned.

...
ignored, for consistency with the generic function.

Value

predict.rendo.ivreg produces a vector of predictions

See also

Examples

data("dataHetIV")

het <- hetErrorsIV(y~X1+X2+P|P|IIV(X1, X2),
                   data = dataHetIV)

#> Residuals were derived by fitting P ~ X1 + X2.

#> The following internal instruments were built: IIV(X1), IIV(X2).

#> Fitting an instrumental variable regression with model y ~ X1 + X2 + P|X1 + X2 + IIV(X1) + IIV(X2).

# returns the fitted values
predict(het)

#>         1         2         3         4         5         6         7         8 
#> 11.488414 10.104987  9.868239 16.167655 15.057722  9.645066 11.047641  8.113010 
#>         9        10        11        12        13        14        15        16 
#> 14.095957 10.667082  7.844275 11.787968  9.298247 12.707932 10.753384 11.040249 
#>        17        18        19        20        21        22        23        24 
#> 10.848032 13.572127 11.557734  9.740487 10.789179 13.415990  9.275542 11.057775 
#>        25        26        27        28        29        30        31        32 
#> 10.060573  9.687347 13.733987 11.595659  7.693705 12.676906 15.164451  8.093210 
#>        33        34        35        36        37        38        39        40 
#> 15.189957 13.059773 11.114011 18.383360 10.927904  9.779328  7.913345  8.837635 
#>        41        42        43        44        45        46        47        48 
#> 10.987300 16.232194 13.114011 14.254982 11.803203 10.987065 17.059866  9.705383 
#>        49        50        51        52        53        54        55        56 
#> 17.065882  8.079849 13.344215 16.583550 12.807899 13.601626  8.804054 12.119791 
#>        57        58        59        60        61        62        63        64 
#> 11.160392 10.448585 10.394923 10.928716 11.400812 14.503386  9.913840 12.124058 
#>        65        66        67        68        69        70        71        72 
#>  9.324392 12.956568 16.433869  7.822448 12.356300  9.911396 10.903316  7.629996 
#>        73        74        75        76        77        78        79        80 
#> 12.729803 10.077350 12.438285  7.849376 15.053450  7.598141 11.456031  8.657266 
#>        81        82        83        84        85        86        87        88 
#> 11.248392 12.260277 12.200587 12.087422 13.416403 15.361157 14.190544 10.735623 
#>        89        90        91        92        93        94        95        96 
#> 10.783919 10.297082 12.581172 15.143217  9.825089 12.375272 11.031382 14.537836 
#>        97        98        99       100       101       102       103       104 
#>  9.114021  9.980382 12.179117 12.430512 13.826565 12.484173 11.796906  8.634895 
#>       105       106       107       108       109       110       111       112 
#> 12.740710 10.108845 10.006441 13.777802 14.904492 12.997446 11.715181  9.923998 
#>       113       114       115       116       117       118       119       120 
#> 12.924090 10.137203 11.028127  9.876902  8.844607  7.570723  8.700243 11.826869 
#>       121       122       123       124       125       126       127       128 
#> 12.256852  9.013435 12.600822 16.303494 11.215715 13.666126 11.005589  9.408278 
#>       129       130       131       132       133       134       135       136 
#> 11.257378 12.711884 12.093820 10.643435 14.957120  9.523462 10.272002 10.813815 
#>       137       138       139       140       141       142       143       144 
#> 12.497235 10.496732 10.444460  8.838153 13.582477  9.958933 12.202183 12.726802 
#>       145       146       147       148       149       150       151       152 
#>  9.671358 10.034273 16.616970 15.413511  9.542632 16.360498 10.267143 10.419821 
#>       153       154       155       156       157       158       159       160 
#> 12.540148  9.458042 11.030358  9.624886 13.843861 11.337664 10.666612  9.453283 
#>       161       162       163       164       165       166       167       168 
#>  8.233029 10.199665 13.273279 10.819070  9.303955  9.966487 10.329131  9.558456 
#>       169       170       171       172       173       174       175       176 
#> 10.879895 13.426967 14.644680 11.492855 12.528412 14.518046  9.534990 15.564505 
#>       177       178       179       180       181       182       183       184 
#> 10.358940 12.063921 10.906308 13.468502  8.145373 12.141805  7.969977  9.016919 
#>       185       186       187       188       189       190       191       192 
#> 11.196661 12.489195 10.453779 16.371716 12.731929 12.425986 12.094949 12.672436 
#>       193       194       195       196       197       198       199       200 
#> 10.047253 10.117466  9.683749 11.709456  9.674507 10.568564 17.222078 11.662183 
#>       201       202       203       204       205       206       207       208 
#> 15.792324 12.812762 10.840782 11.481264 15.946297 15.318448 11.244972 11.497730 
#>       209       210       211       212       213       214       215       216 
#> 12.002922 12.157960 11.654450 11.073665 10.827740 17.216864 10.599455 10.260312 
#>       217       218       219       220       221       222       223       224 
#> 10.694231 14.034075 11.450102 15.191218 10.925989 11.431989 10.730233 14.439873 
#>       225       226       227       228       229       230       231       232 
#> 13.591365 12.619171  8.601728 10.068122 11.770011 12.139846 17.768238 12.883466 
#>       233       234       235       236       237       238       239       240 
#> 10.061657 15.894587 11.554545  9.654331 10.834922  9.966529 10.132449 12.405983 
#>       241       242       243       244       245       246       247       248 
#> 11.099353 12.902204 14.585697 13.262150 10.556738  9.696621  8.582184 13.268501 
#>       249       250       251       252       253       254       255       256 
#> 14.025464  8.872162 12.741120 12.952701 15.162062 13.164627 12.028720  9.438357 
#>       257       258       259       260       261       262       263       264 
#>  9.974333 14.618089  8.896448 10.958096 10.312680  8.735707 12.917959 12.199876 
#>       265       266       267       268       269       270       271       272 
#>  8.721810  9.511350 10.319557 11.078709  9.837703 10.657419 10.875698 11.361677 
#>       273       274       275       276       277       278       279       280 
#>  9.229096  8.323866  9.529284  7.960118 12.302094 12.849339 10.039855  9.893045 
#>       281       282       283       284       285       286       287       288 
#> 10.998065 10.735169 12.694549 11.793333 11.399586  9.130010 10.067800  8.550189 
#>       289       290       291       292       293       294       295       296 
#> 10.239074 10.262217  9.948424  8.966543 12.621260 11.516364 10.343915 15.714455 
#>       297       298       299       300       301       302       303       304 
#>  9.005784 14.274528 12.820822 20.245596 13.428886  8.969759 10.133955 12.523409 
#>       305       306       307       308       309       310       311       312 
#> 11.583233  9.639942 11.618050 10.641678  5.989232 13.347378 11.061594 12.111394 
#>       313       314       315       316       317       318       319       320 
#>  8.664211 12.330155  9.061066 16.547620 12.380261 12.051933  8.061123 10.319070 
#>       321       322       323       324       325       326       327       328 
#>  8.316406 11.557997 17.711060  9.973856 10.558318 10.786511 14.092389  8.560539 
#>       329       330       331       332       333       334       335       336 
#> 10.056822 13.365844 11.860003 10.420347 12.566061 12.938768 16.979705  8.892038 
#>       337       338       339       340       341       342       343       344 
#> 10.642047 11.867924 11.672721  8.508647 12.244000 11.547989 14.654637 11.164346 
#>       345       346       347       348       349       350       351       352 
#> 13.693706  9.642430  9.575679  7.511681  9.436615  9.028774  9.216501 10.596919 
#>       353       354       355       356       357       358       359       360 
#>  7.372034 10.194215 11.094451 12.961891 14.682328 16.263590 11.647265 11.031756 
#>       361       362       363       364       365       366       367       368 
#> 14.410674 11.275330 12.742846 11.129570  9.666207 10.485057 13.844134 10.184416 
#>       369       370       371       372       373       374       375       376 
#>  8.174713 12.679358 13.484689 14.312685 10.028325 17.311601  9.958118 11.813851 
#>       377       378       379       380       381       382       383       384 
#> 14.099500  9.430846 11.737476 12.619653 10.359115  9.964535 12.711067 13.352618 
#>       385       386       387       388       389       390       391       392 
#>  9.171458 11.887977  8.717368 11.523370 10.751835  8.221266 15.018678  9.717107 
#>       393       394       395       396       397       398       399       400 
#> 13.984369  9.984941 13.971495 13.853195 14.121649 13.199331 14.631939 12.190481 
#>       401       402       403       404       405       406       407       408 
#> 13.101325 15.698613 13.499174 10.809441 11.662515  9.532100 12.239978 10.092818 
#>       409       410       411       412       413       414       415       416 
#>  8.170140 18.878817 10.637617 14.095810 15.899169 10.664850 13.931415 12.862135 
#>       417       418       419       420       421       422       423       424 
#> 10.931908  7.203161 13.400412  6.395745 13.101571 11.530659 11.526883 12.803542 
#>       425       426       427       428       429       430       431       432 
#> 17.967806 11.969863 14.198836 13.226804 11.874011  8.463549 11.439406  9.809295 
#>       433       434       435       436       437       438       439       440 
#> 15.005549 15.168396 13.401005 12.808791 17.556833 13.724995 12.489943 10.340521 
#>       441       442       443       444       445       446       447       448 
#> 11.936695  9.916301 11.831138 10.252662  9.956337 12.664914 12.257070  9.877374 
#>       449       450       451       452       453       454       455       456 
#> 11.911392  9.739503 10.236494 12.669481 10.292819 11.999045  8.750019 12.230524 
#>       457       458       459       460       461       462       463       464 
#> 11.163098 10.529353 12.471792 13.813604 11.882893 10.922646  7.471332 10.527121 
#>       465       466       467       468       469       470       471       472 
#> 15.350004 10.658043 13.288651 10.731329  8.686081  8.572292 13.329978 13.703183 
#>       473       474       475       476       477       478       479       480 
#> 10.725692 14.890133 10.997232 12.284702 11.349476 10.018082 10.645471  7.659877 
#>       481       482       483       484       485       486       487       488 
#> 11.618525 10.262834 13.850462 11.198196  9.452749 12.616073  6.925517  8.869319 
#>       489       490       491       492       493       494       495       496 
#> 13.298613 11.072130 11.070803 15.406291 12.866707 10.250465 12.489433 13.438829 
#>       497       498       499       500       501       502       503       504 
#> 14.519074 12.598710 10.882239 11.658819  9.497677  9.913589 12.120173 10.864598 
#>       505       506       507       508       509       510       511       512 
#> 14.077855 10.055754  8.956518 12.625671 16.726044 11.517751  8.839341 16.502524 
#>       513       514       515       516       517       518       519       520 
#>  7.828101 10.663267 11.447857 12.272952 15.287301 11.099420 10.543525  8.620030 
#>       521       522       523       524       525       526       527       528 
#> 13.441243 10.296706 13.438383 11.652692 14.146472  8.753625  8.880525 12.678426 
#>       529       530       531       532       533       534       535       536 
#> 10.591056  7.646317 14.306897  9.017930 11.152254  8.847702 10.527947 13.044588 
#>       537       538       539       540       541       542       543       544 
#> 10.186531 16.004425 11.405579 12.249574 11.391920 14.995059 11.721666 12.247891 
#>       545       546       547       548       549       550       551       552 
#>  9.995712 10.226415 11.127879 10.777191 15.543019 12.336436 15.235404 11.713369 
#>       553       554       555       556       557       558       559       560 
#> 11.729563  8.890726  7.846143 10.208249 13.355123 13.205132  8.902558 10.435677 
#>       561       562       563       564       565       566       567       568 
#> 10.601967 12.188498 13.963590 11.366153  8.898706  9.357149 11.305473 10.380645 
#>       569       570       571       572       573       574       575       576 
#>  9.393479 16.492916 12.282147  9.823955 22.316353 10.151192 10.842783 12.020279 
#>       577       578       579       580       581       582       583       584 
#> 18.481135 11.527624 10.988648 11.768125  9.186809 14.981205 13.517956 11.421378 
#>       585       586       587       588       589       590       591       592 
#> 16.547323 13.150787 16.101354 10.935881  9.474757 10.856385 11.718190 10.498210 
#>       593       594       595       596       597       598       599       600 
#> 16.203960 11.157317 10.207865 10.377503  8.223835  9.590725 11.394646 13.589335 
#>       601       602       603       604       605       606       607       608 
#> 11.384540 12.919403 11.654310  9.234352 14.621953 12.730863 12.890822  7.939080 
#>       609       610       611       612       613       614       615       616 
#> 10.328507 10.462591 16.959285 12.038311  8.979801  9.742460 10.528913  9.047886 
#>       617       618       619       620       621       622       623       624 
#> 13.035237 11.201601 13.164705 11.370747  8.031804  9.459621  8.955716  9.582711 
#>       625       626       627       628       629       630       631       632 
#> 11.249870  8.209755 12.821970 10.162444 11.411881  8.237450 10.344351 11.193522 
#>       633       634       635       636       637       638       639       640 
#>  8.704349 13.903616 11.763343 11.735146 12.336981 10.081926 13.661795 12.997171 
#>       641       642       643       644       645       646       647       648 
#>  9.715641 13.739505 11.488286 12.218104 10.518936 13.181519 13.315181 14.708921 
#>       649       650       651       652       653       654       655       656 
#> 15.178470 15.767543 10.619653 16.378278 10.807327 12.133631 10.276370 12.629004 
#>       657       658       659       660       661       662       663       664 
#> 11.840158 14.987228  8.893329 12.068058 12.716529 10.757174 10.877743 11.579511 
#>       665       666       667       668       669       670       671       672 
#> 13.974714  9.153852 13.105104  9.652917  9.175845 10.911564  7.947299 11.901930 
#>       673       674       675       676       677       678       679       680 
#> 10.173406 11.165300 10.749707 16.968535 11.248890  9.052048 10.443567 16.194029 
#>       681       682       683       684       685       686       687       688 
#> 10.487744 11.535851 11.420792 10.767990  8.044159 11.569285 10.326365  8.244350 
#>       689       690       691       692       693       694       695       696 
#> 11.216206 15.241220  7.852972 13.790203  9.468100  8.689885  6.626593  9.761103 
#>       697       698       699       700       701       702       703       704 
#> 10.493315 12.471784  9.199911  7.424702 11.235793 17.085893 12.882368  9.465908 
#>       705       706       707       708       709       710       711       712 
#> 12.307457 10.517812 11.898105 14.508496 15.299994 11.480027 12.026689 10.288890 
#>       713       714       715       716       717       718       719       720 
#> 13.542553 10.201648  9.009970 11.249094  9.966775 10.341881 11.348503 11.681262 
#>       721       722       723       724       725       726       727       728 
#> 11.154505  8.080566 11.769764 11.653404 10.385184  9.983176 10.080631 10.975140 
#>       729       730       731       732       733       734       735       736 
#>  8.698804 12.885358 10.299992 13.850162 14.225932 11.680382 15.109884 10.427776 
#>       737       738       739       740       741       742       743       744 
#> 11.973263  9.734357 17.484363 11.314792 10.848436 14.005980 14.268524  9.907707 
#>       745       746       747       748       749       750       751       752 
#> 11.276092 13.757700 11.233185 16.182676 11.993008 15.803925 10.425460 12.754142 
#>       753       754       755       756       757       758       759       760 
#> 11.568621 11.289419  7.976021 10.818054  7.826632 10.908791  8.912200 11.365679 
#>       761       762       763       764       765       766       767       768 
#> 12.612939 14.738959 14.530011  9.773565  8.297869 10.967030 12.902512 12.517922 
#>       769       770       771       772       773       774       775       776 
#> 13.820710 13.160077 12.361582  8.539792 11.567570 15.152413 14.942718 12.694720 
#>       777       778       779       780       781       782       783       784 
#> 10.356747  9.872909 10.973509 11.811580 11.066243  8.348377 12.599308 11.218486 
#>       785       786       787       788       789       790       791       792 
#> 13.223630 11.914564 10.998076 10.356344 12.577086 15.037616 10.648023  8.309557 
#>       793       794       795       796       797       798       799       800 
#> 12.229942 10.133165 12.160538 11.120815 13.289939 11.547591  9.595115 10.712183 
#>       801       802       803       804       805       806       807       808 
#> 13.527474 14.995016  8.741171 10.631111  8.433745 10.411752  9.859355 12.300645 
#>       809       810       811       812       813       814       815       816 
#> 10.347866 10.270553 10.765599 12.095920 13.586450  9.648867 13.263300 11.300661 
#>       817       818       819       820       821       822       823       824 
#>  7.090026 10.156888  9.888123  9.237768 14.378542 10.626929 11.525906  8.984630 
#>       825       826       827       828       829       830       831       832 
#> 11.628893 12.824688 12.318835 10.967592 12.618385 12.063286 10.590646 10.987780 
#>       833       834       835       836       837       838       839       840 
#> 12.762591 17.626200 10.914013 11.255071 13.129913  8.378142 13.605991 14.278746 
#>       841       842       843       844       845       846       847       848 
#>  9.883469 13.648337 12.106948 12.594701 12.208701  8.398546 10.688657 15.580890 
#>       849       850       851       852       853       854       855       856 
#> 12.188912 10.942289 10.173354 20.853598 10.480238 13.055192  9.696938 10.187744 
#>       857       858       859       860       861       862       863       864 
#> 11.841759 15.341202 10.747961 11.093645 14.545802 10.466433 10.036355 16.464033 
#>       865       866       867       868       869       870       871       872 
#>  9.115972 10.438540 10.781527 11.804353  9.263308 11.743178  9.981765  8.497238 
#>       873       874       875       876       877       878       879       880 
#> 10.773031  9.466994 14.267221 10.432148 13.109037  9.914138 16.107401 10.720291 
#>       881       882       883       884       885       886       887       888 
#> 17.278772  9.606643 10.258454 12.885747 10.340794 10.246076 13.706846 11.936536 
#>       889       890       891       892       893       894       895       896 
#> 11.098008  8.074340 13.896951 11.697264  9.768938 11.421751 11.253194 10.691360 
#>       897       898       899       900       901       902       903       904 
#> 11.806415 14.879508 10.667275 11.529824 11.201316 12.724868 10.491746  8.511035 
#>       905       906       907       908       909       910       911       912 
#> 12.160990 12.851695 12.783981 11.283107 11.509564 11.537171  9.972122 14.074464 
#>       913       914       915       916       917       918       919       920 
#> 11.365005 11.525029 10.930305 11.813570 12.767400 11.722084 15.838669 16.478438 
#>       921       922       923       924       925       926       927       928 
#> 13.252264  9.433304 10.106572 10.176174  8.106843 11.247893 13.892151 10.097836 
#>       929       930       931       932       933       934       935       936 
#> 14.940278 10.659740 11.672051 13.382560  8.645321 12.873155 12.749897 10.160781 
#>       937       938       939       940       941       942       943       944 
#>  9.398653  9.774867  9.817423 10.005027 10.999498 11.979530  7.915176 12.768998 
#>       945       946       947       948       949       950       951       952 
#> 10.488707 12.540976 12.251320 11.119760 12.120999 10.225851 10.862523 15.865852 
#>       953       954       955       956       957       958       959       960 
#> 13.673927 12.726058 12.340799 10.097656  6.771180 11.419319 10.058605  9.845444 
#>       961       962       963       964       965       966       967       968 
#>  9.914348  9.734559  8.917940 10.184433 10.046211 11.011765 11.423217 14.598045 
#>       969       970       971       972       973       974       975       976 
#> 10.542206  9.335286 12.626621  9.567445 13.288440 15.778763 10.375522 12.730502 
#>       977       978       979       980       981       982       983       984 
#> 11.948534 15.765055 10.505633  9.970903  9.783146 15.194930 11.471125 10.460860 
#>       985       986       987       988       989       990       991       992 
#>  9.652295 10.468997 13.712353 13.048175 14.217269  9.238340 11.260882 14.513086 
#>       993       994       995       996       997       998       999      1000 
#>  8.763837 14.762790 12.625756 12.090010 11.344512 10.497013  7.883945  9.730532 
#>      1001      1002      1003      1004      1005      1006      1007      1008 
#> 12.010933 14.094875  6.954729 12.360954 13.388845 10.696235 13.677011 12.113786 
#>      1009      1010      1011      1012      1013      1014      1015      1016 
#> 10.929501  9.385025 10.688741 13.380863 10.811419 13.679791 15.241327 10.606383 
#>      1017      1018      1019      1020      1021      1022      1023      1024 
#>  8.354923 12.409974 12.210388  7.877099 13.118557 10.405984 13.045831  8.022339 
#>      1025      1026      1027      1028      1029      1030      1031      1032 
#> 13.120421  8.794381 11.149425 13.014332  8.871265  9.330794 15.769763 13.386197 
#>      1033      1034      1035      1036      1037      1038      1039      1040 
#>  9.635409 10.441513 15.503317 16.007528 10.412043 13.967085 17.034133 17.942625 
#>      1041      1042      1043      1044      1045      1046      1047      1048 
#> 10.039924 18.261133 14.759887 12.020510 11.182414 15.839595 12.037968 10.949866 
#>      1049      1050      1051      1052      1053      1054      1055      1056 
#> 11.548660  9.780964 12.302143 14.966466 13.756496  9.726200 15.124725 12.858110 
#>      1057      1058      1059      1060      1061      1062      1063      1064 
#> 13.627975  7.688468 12.839888 11.299007  7.161176 13.327777 12.714263 14.269150 
#>      1065      1066      1067      1068      1069      1070      1071      1072 
#> 13.650278 12.316745 11.571852 17.032302 11.404110 13.183254  8.691474 12.994640 
#>      1073      1074      1075      1076      1077      1078      1079      1080 
#>  9.815941 11.898259 11.899588 11.271228  9.693632 11.174386  9.553557 17.262860 
#>      1081      1082      1083      1084      1085      1086      1087      1088 
#> 12.905858 13.114591 11.065306 13.048708 13.814929 11.334485 13.590759 11.683968 
#>      1089      1090      1091      1092      1093      1094      1095      1096 
#> 12.191468 11.917449 12.784065 12.235923 12.712023 10.752550  8.857994 11.204717 
#>      1097      1098      1099      1100      1101      1102      1103      1104 
#> 17.450570 14.944968 12.968737  9.640275  9.932982 11.229664 11.679482 10.885054 
#>      1105      1106      1107      1108      1109      1110      1111      1112 
#> 11.731337 11.048507 10.772824 10.650380 11.208232 12.307025  7.705661 12.068414 
#>      1113      1114      1115      1116      1117      1118      1119      1120 
#> 10.308622 12.303087 11.362451  9.801319  8.198598 14.532618 10.584528  9.563587 
#>      1121      1122      1123      1124      1125      1126      1127      1128 
#>  8.859248 12.069012  8.596336 11.574946  8.514042 11.306029 12.076050 10.214746 
#>      1129      1130      1131      1132      1133      1134      1135      1136 
#> 11.878503  9.367723 13.116139  9.249851 10.228356 13.688536  9.976764 12.056459 
#>      1137      1138      1139      1140      1141      1142      1143      1144 
#>  9.773701  9.669701  7.510456  8.629511 13.681932  9.446239 10.215473  7.977820 
#>      1145      1146      1147      1148      1149      1150      1151      1152 
#> 10.576777 14.653859  8.546692 10.185808 11.102779 10.937263 11.448507  9.669539 
#>      1153      1154      1155      1156      1157      1158      1159      1160 
#> 14.244065 10.639690 15.513071  9.165305  8.687525  9.489828  6.579793  8.299117 
#>      1161      1162      1163      1164      1165      1166      1167      1168 
#>  9.705741  9.590142 12.074980 15.852891 11.280288 15.618429 15.025170 10.433921 
#>      1169      1170      1171      1172      1173      1174      1175      1176 
#> 11.449146 12.552885 10.180640 11.953687 13.003789  7.279290  9.394165  8.740074 
#>      1177      1178      1179      1180      1181      1182      1183      1184 
#> 12.968381 14.378135 10.059836 10.332839 15.080944 14.304501 10.689108 12.657687 
#>      1185      1186      1187      1188      1189      1190      1191      1192 
#> 10.703428 12.561601  8.443092  8.685303 12.113413  7.314479  9.715082  8.516891 
#>      1193      1194      1195      1196      1197      1198      1199      1200 
#> 10.935277 13.652462  8.036820 12.152756 13.513928  9.769708  8.611542 13.479724 
#>      1201      1202      1203      1204      1205      1206      1207      1208 
#> 11.105438 12.031762 10.632057 10.241500 10.186716 11.642273 10.363940 10.938721 
#>      1209      1210      1211      1212      1213      1214      1215      1216 
#> 14.715384 18.446836 12.557782  8.939255  7.917629 12.340603 13.363155 12.192999 
#>      1217      1218      1219      1220      1221      1222      1223      1224 
#>  9.861509 13.006771 13.342419 11.123530 10.671557 11.329552 10.323274 15.779273 
#>      1225      1226      1227      1228      1229      1230      1231      1232 
#>  8.951678 12.370133  8.842114 14.168024 11.362303 16.154938 13.171801 13.112676 
#>      1233      1234      1235      1236      1237      1238      1239      1240 
#>  9.927312  8.741620 14.475447 10.820491  8.524060 11.871517 10.036491 10.591637 
#>      1241      1242      1243      1244      1245      1246      1247      1248 
#> 11.226138 13.426096 11.991727 13.221305 13.350238  8.169740  8.229745  9.362050 
#>      1249      1250      1251      1252      1253      1254      1255      1256 
#> 10.564108 12.634757 10.144263 14.371781 11.511392 12.145242 11.046283 10.211171 
#>      1257      1258      1259      1260      1261      1262      1263      1264 
#>  9.075305  9.056230 10.651785 11.341167 10.825592  8.228570  9.961063 11.262311 
#>      1265      1266      1267      1268      1269      1270      1271      1272 
#> 10.690320  9.651463  9.386102  9.597174 15.409447 11.770914 14.790528  8.814796 
#>      1273      1274      1275      1276      1277      1278      1279      1280 
#> 12.537514  9.863198 10.903911 11.167675 10.218452 11.922085 11.908153 13.618174 
#>      1281      1282      1283      1284      1285      1286      1287      1288 
#> 11.898107 10.424739  9.588580 14.127666 13.116931 10.489216 12.835674  9.174849 
#>      1289      1290      1291      1292      1293      1294      1295      1296 
#> 12.740384 11.390630 18.229739 12.719967 10.874722 11.490796 10.787585 11.107315 
#>      1297      1298      1299      1300      1301      1302      1303      1304 
#> 10.517956 13.663858 13.531651 15.155031 10.904482 11.935388  7.552677 13.787999 
#>      1305      1306      1307      1308      1309      1310      1311      1312 
#>  9.094980 12.218222 15.648696  9.038576 10.653550 12.093256  9.425773 11.651209 
#>      1313      1314      1315      1316      1317      1318      1319      1320 
#>  9.762809  9.759038 13.265267  7.134841 13.348278  9.475150 11.364039 14.117831 
#>      1321      1322      1323      1324      1325      1326      1327      1328 
#> 14.804479 11.560426  9.144730  9.140054  7.712403 13.135075 10.576856 14.386735 
#>      1329      1330      1331      1332      1333      1334      1335      1336 
#> 12.255158 11.411801  9.502942 10.813347 10.050122  9.953054  8.802643 14.058671 
#>      1337      1338      1339      1340      1341      1342      1343      1344 
#> 14.477793  9.925049 11.263615  8.696013 13.304971  9.425371 12.046592 11.084288 
#>      1345      1346      1347      1348      1349      1350      1351      1352 
#> 11.306323  7.471976 11.114433 12.229875  8.510389 10.043850 14.577655 16.154659 
#>      1353      1354      1355      1356      1357      1358      1359      1360 
#> 13.618214  9.082132  9.813947 13.788423  8.674427 11.686166 10.862846 12.262388 
#>      1361      1362      1363      1364      1365      1366      1367      1368 
#> 11.455127 10.121248  8.961787 14.381222 13.639971 11.776766 11.741251  8.785455 
#>      1369      1370      1371      1372      1373      1374      1375      1376 
#>  8.781079 12.747459 14.232772 11.318419 11.375274 10.821878 10.234665 11.617724 
#>      1377      1378      1379      1380      1381      1382      1383      1384 
#>  9.521688 12.120388 11.343148 14.006969 10.897870  9.726865 14.746791 10.058963 
#>      1385      1386      1387      1388      1389      1390      1391      1392 
#> 14.886772 12.665162  6.965914 12.192965 11.345768 11.680914 11.031467 11.731374 
#>      1393      1394      1395      1396      1397      1398      1399      1400 
#> 11.940632 16.368606 15.487614  8.077392 14.623749 11.721476  9.919866 10.401910 
#>      1401      1402      1403      1404      1405      1406      1407      1408 
#> 10.966692 11.462560 10.546740  7.897536 11.085621 13.184902 12.740847 13.279416 
#>      1409      1410      1411      1412      1413      1414      1415      1416 
#> 10.548103 13.925150 12.418237 11.982613 10.132794 12.906432 12.659394 13.694596 
#>      1417      1418      1419      1420      1421      1422      1423      1424 
#>  8.510127 12.329944 11.520536 15.945061 13.236015  9.065885  9.167016 10.112208 
#>      1425      1426      1427      1428      1429      1430      1431      1432 
#> 12.349388 10.647739 12.019689 10.255313 10.071320 15.333542  8.463115 16.825727 
#>      1433      1434      1435      1436      1437      1438      1439      1440 
#>  9.306361  9.095136 12.858506 13.683276 16.588172 11.396045  8.009309  9.706388 
#>      1441      1442      1443      1444      1445      1446      1447      1448 
#>  9.965176 13.709855 12.083566 13.138398  9.568171 13.152278 11.939447  8.518596 
#>      1449      1450      1451      1452      1453      1454      1455      1456 
#> 11.240802 12.826989 17.056223 13.136214 13.131584 12.216286 14.101383 13.502761 
#>      1457      1458      1459      1460      1461      1462      1463      1464 
#> 11.565806 12.804571 12.615362  8.394694  8.308615 11.215896 13.790701 13.124298 
#>      1465      1466      1467      1468      1469      1470      1471      1472 
#> 10.436670 14.467581 11.019898 12.798059  7.077251  9.984754 10.234856 10.547401 
#>      1473      1474      1475      1476      1477      1478      1479      1480 
#> 11.329246 10.946163 10.438041 11.400698 12.242787 12.053119 10.327577  8.754293 
#>      1481      1482      1483      1484      1485      1486      1487      1488 
#> 14.937823 14.509135 14.769434 14.069653  9.955801  9.037106 11.427519 14.874069 
#>      1489      1490      1491      1492      1493      1494      1495      1496 
#> 11.337956  9.573274 11.950599 11.761570  8.264351 12.961341 11.507318 12.956727 
#>      1497      1498      1499      1500      1501      1502      1503      1504 
#> 12.138288 10.314565 11.373407 10.365931 10.492950 10.542305 14.371007 12.461259 
#>      1505      1506      1507      1508      1509      1510      1511      1512 
#> 11.370065 12.650278 16.960614 14.419461 12.434567  7.897011 10.925446  8.988247 
#>      1513      1514      1515      1516      1517      1518      1519      1520 
#> 12.983572 12.334872 11.511181 11.545757 11.001903  9.833683 14.861558 10.868262 
#>      1521      1522      1523      1524      1525      1526      1527      1528 
#> 14.561647 13.711988 11.210012 10.609519  7.988503  9.983879 13.918939 10.887756 
#>      1529      1530      1531      1532      1533      1534      1535      1536 
#>  8.970499 13.790256 11.402636 10.875412 10.559998  9.701509  8.878296  9.888495 
#>      1537      1538      1539      1540      1541      1542      1543      1544 
#> 10.285005  9.142959  9.372128 12.632306  7.784809 13.846977  8.546057 14.146392 
#>      1545      1546      1547      1548      1549      1550      1551      1552 
#>  9.424704 13.940948 12.972417 15.155825 11.528303  9.178197 11.515637 10.878953 
#>      1553      1554      1555      1556      1557      1558      1559      1560 
#>  9.590398 11.056156  9.994912 11.238575 11.142551  7.418399 11.762558 12.826471 
#>      1561      1562      1563      1564      1565      1566      1567      1568 
#> 11.884345 10.343912 11.271172 10.790984 12.596250 12.964340 11.583993 14.051122 
#>      1569      1570      1571      1572      1573      1574      1575      1576 
#> 12.228705 10.134893 12.771162 14.038652 10.648701 10.929090  9.699223  9.588448 
#>      1577      1578      1579      1580      1581      1582      1583      1584 
#> 11.313527 12.802774  9.668242 10.977112 11.340709 11.476925  9.522497 12.192555 
#>      1585      1586      1587      1588      1589      1590      1591      1592 
#> 10.000818 11.713373 12.693229 11.371568 13.219990  9.449550 11.113344 11.997679 
#>      1593      1594      1595      1596      1597      1598      1599      1600 
#> 13.422770 14.567677  9.052992  6.617059  8.943271 15.837111  9.097749 14.425448 
#>      1601      1602      1603      1604      1605      1606      1607      1608 
#>  9.876550 12.689501  9.503424 12.870852 12.780561  8.975802 10.752430 11.021831 
#>      1609      1610      1611      1612      1613      1614      1615      1616 
#> 15.765750  8.596543  9.278908  9.205141 17.729405 11.893125 11.915572 11.020275 
#>      1617      1618      1619      1620      1621      1622      1623      1624 
#> 10.888039  9.867924 10.019264  7.865245 12.518504  9.638605 15.171058 11.491974 
#>      1625      1626      1627      1628      1629      1630      1631      1632 
#> 17.419174 11.997671  9.171075 10.298928 14.195402 11.565567  9.949055 14.051955 
#>      1633      1634      1635      1636      1637      1638      1639      1640 
#> 10.978495 10.947415 10.067743 14.059157  9.629149 10.977859 11.936563 14.749236 
#>      1641      1642      1643      1644      1645      1646      1647      1648 
#> 11.315576 11.062949 12.193188  9.916276 18.605195 11.153544  9.532134 11.962962 
#>      1649      1650      1651      1652      1653      1654      1655      1656 
#> 15.129031 16.219436 13.104939 11.960350  8.617555  9.950735 13.444715  9.027869 
#>      1657      1658      1659      1660      1661      1662      1663      1664 
#> 13.900286 11.392364 11.912220 11.956411  9.462588 15.855951  9.900103 12.510544 
#>      1665      1666      1667      1668      1669      1670      1671      1672 
#> 10.429691 11.490541  8.217272 12.219979  9.393818 13.724900  6.838977 12.238296 
#>      1673      1674      1675      1676      1677      1678      1679      1680 
#> 13.299004  9.323221  9.891062 10.617880 11.147851 10.678150  7.466330 12.799738 
#>      1681      1682      1683      1684      1685      1686      1687      1688 
#> 10.772792 17.133316  7.392874 11.911090 13.944880 12.441575 13.939621 11.265043 
#>      1689      1690      1691      1692      1693      1694      1695      1696 
#>  9.042538  9.729194  8.843779 11.914310 13.095720 12.202341 12.035404 13.501542 
#>      1697      1698      1699      1700      1701      1702      1703      1704 
#>  9.799843 10.218241 13.101077 14.846113 11.416833  9.975820 11.838297  9.607636 
#>      1705      1706      1707      1708      1709      1710      1711      1712 
#> 13.509059  6.727635 10.624886 13.281298 11.640595  9.250054 12.631581 11.210803 
#>      1713      1714      1715      1716      1717      1718      1719      1720 
#> 12.264975 10.783148  9.596028 11.675023 12.296858 17.305304 12.959866 12.541920 
#>      1721      1722      1723      1724      1725      1726      1727      1728 
#> 10.756817 14.125016  8.851722 13.705833 11.807999  8.914230  9.644488 11.820987 
#>      1729      1730      1731      1732      1733      1734      1735      1736 
#> 10.336887  9.684085 12.518066 14.467419 13.216141  9.046262 11.645401 14.836512 
#>      1737      1738      1739      1740      1741      1742      1743      1744 
#> 10.997234 14.881813 16.886397 13.620330 11.799440 13.387422 12.188826 12.599170 
#>      1745      1746      1747      1748      1749      1750      1751      1752 
#> 15.792360 12.002241 13.237667 11.140845 11.850416 11.218977 12.606818 10.227230 
#>      1753      1754      1755      1756      1757      1758      1759      1760 
#> 11.404385 11.752881 11.064165  9.581585  8.644424 11.322219 14.162835 12.654192 
#>      1761      1762      1763      1764      1765      1766      1767      1768 
#> 13.202249 15.774286 10.054872 11.765134 11.837036 11.157740 13.199693 11.161379 
#>      1769      1770      1771      1772      1773      1774      1775      1776 
#>  7.123563 10.676267  9.730931 11.061602 10.511998 17.621800  8.821787 11.587064 
#>      1777      1778      1779      1780      1781      1782      1783      1784 
#> 11.953417 15.864804 10.750592 11.709945 12.780739  9.512758 13.289768 11.708992 
#>      1785      1786      1787      1788      1789      1790      1791      1792 
#> 13.121412 10.324196 14.049629 12.164315 16.030086  7.917585 12.219732 10.685239 
#>      1793      1794      1795      1796      1797      1798      1799      1800 
#> 11.520171 14.039720 10.776768 14.069069 13.101538 12.460562 13.842471 12.890414 
#>      1801      1802      1803      1804      1805      1806      1807      1808 
#> 11.168710 13.467599 12.235402 13.453067 14.199869 13.842753 10.640571 15.034548 
#>      1809      1810      1811      1812      1813      1814      1815      1816 
#> 11.655756  7.918216 11.071260 13.230112 11.864078 12.367505 12.601645  7.085324 
#>      1817      1818      1819      1820      1821      1822      1823      1824 
#> 10.457907 12.516402  9.455033 12.063493 13.776456 11.540719 10.466553  8.187029 
#>      1825      1826      1827      1828      1829      1830      1831      1832 
#> 10.907401 10.309094 11.051634 13.214619 10.967355 11.145229 11.612225  9.683478 
#>      1833      1834      1835      1836      1837      1838      1839      1840 
#> 12.163266 13.461996  9.283259  9.245935 12.482713 14.819299 10.304900 13.678879 
#>      1841      1842      1843      1844      1845      1846      1847      1848 
#> 10.654030  8.515249 11.001592 12.073394 12.240013 15.281405 11.765048  8.648319 
#>      1849      1850      1851      1852      1853      1854      1855      1856 
#> 11.067193  8.476071 15.639144 12.104613  7.785039 10.671184 11.880356  7.721037 
#>      1857      1858      1859      1860      1861      1862      1863      1864 
#> 11.229948 10.655765  9.957670  8.812905 11.602207 13.187020 11.218231 11.558186 
#>      1865      1866      1867      1868      1869      1870      1871      1872 
#> 11.169835  9.558572  8.899736  9.202519 11.286274 15.237308 13.894139 16.073535 
#>      1873      1874      1875      1876      1877      1878      1879      1880 
#> 11.887587 10.434136 12.445999 11.460808 15.957627 12.332092  8.654638 13.062635 
#>      1881      1882      1883      1884      1885      1886      1887      1888 
#> 10.457728 13.514644 14.995673  9.529041 12.269654 10.474588  9.558480 10.023934 
#>      1889      1890      1891      1892      1893      1894      1895      1896 
#> 12.636320  9.135003 11.850924 18.146063  9.160692 11.014620 10.391883 10.233019 
#>      1897      1898      1899      1900      1901      1902      1903      1904 
#>  9.848844 15.080768 11.369177 10.887450 13.085520 11.685837  7.854763  9.695151 
#>      1905      1906      1907      1908      1909      1910      1911      1912 
#> 12.826645 11.717419 14.362670 10.944054 10.580480 12.193189  9.869484 15.025035 
#>      1913      1914      1915      1916      1917      1918      1919      1920 
#> 10.723860 12.321457 12.510883 16.298917  7.046960 16.055026 18.313880 14.411539 
#>      1921      1922      1923      1924      1925      1926      1927      1928 
#> 12.403943 12.926121 12.442640 12.089670 15.464066 12.741578  9.660086 11.972957 
#>      1929      1930      1931      1932      1933      1934      1935      1936 
#> 12.063182 14.834744 16.871569 16.248789 10.242572  9.630571 16.608699 12.447994 
#>      1937      1938      1939      1940      1941      1942      1943      1944 
#>  9.162277 10.131413 14.652507 10.973295 10.769736  8.680971 14.154474  8.224881 
#>      1945      1946      1947      1948      1949      1950      1951      1952 
#>  9.314591 14.882396 20.608845 12.247687  8.865908  9.457433 11.920658 11.136320 
#>      1953      1954      1955      1956      1957      1958      1959      1960 
#> 12.770873 10.990761  9.151930 10.884579 14.281104  9.923536 12.670694 11.079147 
#>      1961      1962      1963      1964      1965      1966      1967      1968 
#>  7.969792  7.351961 11.676120 14.457413  9.490529  9.512726 10.263326 14.035369 
#>      1969      1970      1971      1972      1973      1974      1975      1976 
#> 15.039961 10.753114 11.434816 12.248503 13.520138  7.585346 10.927737  7.752410 
#>      1977      1978      1979      1980      1981      1982      1983      1984 
#> 20.669643 13.254592 14.474544 14.026136  8.471305 13.631915 17.632777 13.266658 
#>      1985      1986      1987      1988      1989      1990      1991      1992 
#>  9.478345 10.602595 16.281994 12.477186 11.476320 12.810918 10.769577 13.568280 
#>      1993      1994      1995      1996      1997      1998      1999      2000 
#>  7.619927 12.899198 12.301878 11.406865 10.114610 10.517013 13.660263 13.542563 
#>      2001      2002      2003      2004      2005      2006      2007      2008 
#> 13.133646  9.502894 11.023741 16.911475 14.412849 14.913717  8.677527 11.904660 
#>      2009      2010      2011      2012      2013      2014      2015      2016 
#> 10.603145  9.765888 14.100090  9.805990 11.572188 12.302024  9.112428 14.435185 
#>      2017      2018      2019      2020      2021      2022      2023      2024 
#> 10.832826 10.906997 11.007561 12.823290 12.484159 11.835559 11.699659  8.716649 
#>      2025      2026      2027      2028      2029      2030      2031      2032 
#>  9.067354 14.230139 10.275830 12.720831 10.486377 14.287162 11.148754 10.917316 
#>      2033      2034      2035      2036      2037      2038      2039      2040 
#> 13.409753 11.935853  9.695350 12.201428 11.398264  8.486817  9.826006 11.950514 
#>      2041      2042      2043      2044      2045      2046      2047      2048 
#> 10.705467 10.170935 10.169811 11.356395 10.731804  8.232287  8.418531  9.997487 
#>      2049      2050      2051      2052      2053      2054      2055      2056 
#> 16.342514 14.080074 14.732672  9.024115 15.357083 12.420729 13.520961 12.698067 
#>      2057      2058      2059      2060      2061      2062      2063      2064 
#> 11.387351 11.103215 10.081946 11.026219 13.837808 15.227253  9.961094 10.668384 
#>      2065      2066      2067      2068      2069      2070      2071      2072 
#> 15.192497 10.798042  9.784749 10.003726 10.472973 12.667998 14.417947  4.732454 
#>      2073      2074      2075      2076      2077      2078      2079      2080 
#> 13.215451 12.296146  9.456944 12.974073  9.794332 12.666486 10.400575 12.911108 
#>      2081      2082      2083      2084      2085      2086      2087      2088 
#> 12.925238  9.657040 12.316967 14.603340 13.357556 10.014089 11.844943 11.377115 
#>      2089      2090      2091      2092      2093      2094      2095      2096 
#> 10.601682  9.466329  9.231892  7.557754 11.377539 14.414346 11.586155 15.393181 
#>      2097      2098      2099      2100      2101      2102      2103      2104 
#> 15.485417 12.239112 11.408245 11.339431 11.061322 12.064500 10.277215  7.204350 
#>      2105      2106      2107      2108      2109      2110      2111      2112 
#>  8.700574 12.598914 10.435762 10.875007 14.083245 12.850240  8.849681  9.292018 
#>      2113      2114      2115      2116      2117      2118      2119      2120 
#>  9.142060 10.711532 11.316947 14.952334 15.683282 11.155977 10.529478 13.953018 
#>      2121      2122      2123      2124      2125      2126      2127      2128 
#> 15.623114  7.602399 14.326293 13.340178 13.753467 15.405531 11.375997 13.617153 
#>      2129      2130      2131      2132      2133      2134      2135      2136 
#> 11.851269 12.711030 10.783028 10.626441 17.074597 12.289984  8.664393  8.417311 
#>      2137      2138      2139      2140      2141      2142      2143      2144 
#> 11.231060  8.127479 12.412314 14.962092 12.497960 10.554477  8.112970 11.007351 
#>      2145      2146      2147      2148      2149      2150      2151      2152 
#>  9.780078 11.001973 14.841894 11.483127 11.325550  8.068653 12.674830  8.914513 
#>      2153      2154      2155      2156      2157      2158      2159      2160 
#> 13.274501 11.590151 11.553723 11.443842 10.904048 15.440011 13.794060 10.997731 
#>      2161      2162      2163      2164      2165      2166      2167      2168 
#>  9.180524 11.823044 10.816549  9.034450 10.522937 11.105821 11.424321 11.445217 
#>      2169      2170      2171      2172      2173      2174      2175      2176 
#> 14.161617 12.719228 10.802771 17.413352 13.022794 12.613706  9.126066 15.097506 
#>      2177      2178      2179      2180      2181      2182      2183      2184 
#> 10.370311 19.547900 14.148911 10.686553  7.840527 11.179985  9.649771 11.210177 
#>      2185      2186      2187      2188      2189      2190      2191      2192 
#>  9.234507 16.627909 11.233542 12.763253  9.676885 15.229323 14.692156  8.800933 
#>      2193      2194      2195      2196      2197      2198      2199      2200 
#>  9.777162 14.303266 10.782563 10.681355  9.846300  9.515539  9.952599 12.976862 
#>      2201      2202      2203      2204      2205      2206      2207      2208 
#> 14.362850 13.864033  8.978767 12.009100  8.587255 10.832768 10.482205 11.623845 
#>      2209      2210      2211      2212      2213      2214      2215      2216 
#> 10.776665 12.788568 14.879289 10.359899 10.267983 10.519209  9.598630 17.205508 
#>      2217      2218      2219      2220      2221      2222      2223      2224 
#> 11.920479 13.139117  9.386433 13.565787 10.631721  8.612338  9.887581  8.798453 
#>      2225      2226      2227      2228      2229      2230      2231      2232 
#> 10.332654 11.879099  9.525420 10.119095 14.168244  9.285708 11.503778  9.456245 
#>      2233      2234      2235      2236      2237      2238      2239      2240 
#> 10.698090 11.454890 12.340882 11.785394 14.252614  9.190887 11.675725  8.894379 
#>      2241      2242      2243      2244      2245      2246      2247      2248 
#> 11.034105 14.733417 14.379661  9.200380 11.147285 12.553342  7.617361 15.562267 
#>      2249      2250      2251      2252      2253      2254      2255      2256 
#> 15.959029  6.142937 11.666410 13.404387  9.829163  9.561652 15.461156 19.062002 
#>      2257      2258      2259      2260      2261      2262      2263      2264 
#> 10.768121 13.591678 16.170876 11.383533 12.566790 11.180118 18.114196 10.409750 
#>      2265      2266      2267      2268      2269      2270      2271      2272 
#>  9.115547 11.263580  7.507703  9.928367 10.841562 12.495038 10.434765 11.440446 
#>      2273      2274      2275      2276      2277      2278      2279      2280 
#> 13.894284 13.616157  9.659981 13.143222 11.590508 12.618697 11.032215 14.265816 
#>      2281      2282      2283      2284      2285      2286      2287      2288 
#> 11.788789 12.870256 10.354783 11.240057 11.904491  9.494532 10.451802  9.839806 
#>      2289      2290      2291      2292      2293      2294      2295      2296 
#> 10.576788 11.542674  9.271781 11.944832 15.104610 13.164051 14.198409  8.844435 
#>      2297      2298      2299      2300      2301      2302      2303      2304 
#> 14.896925 15.818672  9.624077 16.028070 10.959854  8.435938 10.530002 15.203361 
#>      2305      2306      2307      2308      2309      2310      2311      2312 
#> 11.638363 12.715314 13.813538  8.082766 11.535678 10.220221 10.692382 14.899389 
#>      2313      2314      2315      2316      2317      2318      2319      2320 
#> 15.081843 15.440999 13.803277 11.012788 11.313551 12.194827  8.869810 11.584857 
#>      2321      2322      2323      2324      2325      2326      2327      2328 
#> 12.156318  8.921109 11.408838 11.348450  7.242437  9.619691 12.391679 10.834292 
#>      2329      2330      2331      2332      2333      2334      2335      2336 
#> 14.813162 14.952979 10.127849 10.925814 11.462556 11.152978 10.605941  8.899065 
#>      2337      2338      2339      2340      2341      2342      2343      2344 
#> 17.633254 10.408878 14.318626 10.366688 12.530911 18.973721 12.927252 11.852445 
#>      2345      2346      2347      2348      2349      2350      2351      2352 
#> 14.085170 14.416438 11.103181 16.199575 11.732956  8.574924  9.438130 13.908893 
#>      2353      2354      2355      2356      2357      2358      2359      2360 
#>  9.993515 14.960794  9.694411 10.119231 12.895985 12.878879  9.965373  9.544966 
#>      2361      2362      2363      2364      2365      2366      2367      2368 
#>  8.739046  9.825924 12.478795  9.884379 12.991927  9.361916  7.882084 15.543015 
#>      2369      2370      2371      2372      2373      2374      2375      2376 
#>  9.356442 14.904898 15.208127 10.248270 11.005873 12.865878 12.555199 10.058387 
#>      2377      2378      2379      2380      2381      2382      2383      2384 
#>  8.555037 10.894693  9.686822 11.529366 12.533062 12.875076  8.367265  8.055687 
#>      2385      2386      2387      2388      2389      2390      2391      2392 
#> 11.878088  9.145667 15.077993 10.544173  9.341307  9.197870 10.853662 14.571007 
#>      2393      2394      2395      2396      2397      2398      2399      2400 
#>  9.682789 12.756079 14.752475 17.707860 12.806805 15.014369 13.397159 13.685721 
#>      2401      2402      2403      2404      2405      2406      2407      2408 
#> 10.912118 15.603578  9.607423  7.528472 10.569107 14.463882 11.595211 12.174783 
#>      2409      2410      2411      2412      2413      2414      2415      2416 
#> 12.781364 13.065083 11.563491 14.643619  8.765968 10.309488 11.771133 15.761077 
#>      2417      2418      2419      2420      2421      2422      2423      2424 
#> 12.544191 13.033256 11.195552 10.277443 11.892317 10.495014 10.380136 17.640858 
#>      2425      2426      2427      2428      2429      2430      2431      2432 
#>  8.519568 14.574879 11.184043 10.720019  9.799204 11.979186 10.127748 13.613023 
#>      2433      2434      2435      2436      2437      2438      2439      2440 
#> 11.574289 11.619009  8.178118  9.807670 10.742571 10.101238 12.592473 18.293699 
#>      2441      2442      2443      2444      2445      2446      2447      2448 
#> 15.299919 13.473909 10.527233  8.933283 11.729186 11.368195 10.586616 10.311929 
#>      2449      2450      2451      2452      2453      2454      2455      2456 
#>  9.193387  9.658438 16.975920 12.828676 10.672542 12.073033 13.554477 12.665847 
#>      2457      2458      2459      2460      2461      2462      2463      2464 
#> 12.323741 12.096459  8.526406 12.178725  8.272851  9.207050 10.933431 10.971570 
#>      2465      2466      2467      2468      2469      2470      2471      2472 
#> 11.036743 13.577208 12.120937  9.098569 12.003012 13.149034 14.086625 11.749734 
#>      2473      2474      2475      2476      2477      2478      2479      2480 
#>  8.933588 13.678517  9.660746 11.244625 11.312906 13.011523 18.360387 10.911289 
#>      2481      2482      2483      2484      2485      2486      2487      2488 
#> 12.627590  9.288236 13.293377  8.872993 10.467840 11.213741 12.045499 10.847518 
#>      2489      2490      2491      2492      2493      2494      2495      2496 
#> 11.458956 12.234495 11.526442 12.630804 13.416240 11.813293 12.724027 10.649379 
#>      2497      2498      2499      2500 
#> 10.273916 13.270894  7.996432  9.426123 

# using the data used for fitting also for predicting,
# correctly results in fitted values
all.equal(predict(het, dataHetIV), fitted(het)) # TRUE

#> [1] TRUE