Summarizing latentIV Model Fits

summary method for a model of class rendo.latent.IV resulting from fitting latentIV

# S3 method for rendo.latent.IV
summary(object, ...)

Arguments

object

an object of class rendo.latent.IV, a result of a call to latentIV.

...

ignored, for consistency with the generic function.

Value

The function summary.rendo.latent.IV computes and returns a list of summary statistics which contains the following components:

coefficients

a px4 matrix with columns for the estimated coefficients, its standard error, the t-statistic and corresponding (two-sided) p-value.

start.params

a named vector with the initial set of parameters used to optimize the log-likelihood function.

names.main.coefs

a vector specifying which coefficients are from the model. For internal usage.

vcov

variance covariance matrix derived from the hessian.

AIC

Akaike's An Information Criterion for the model fitted on the provided data.

BIC

Schwarz's Bayesian Criterion for the model fitted on the provided data.

KKT1

first Kuhn, Karush, Tucker optimality condition as returned by optimx.

KKT2

second Kuhn, Karush, Tucker optimality condition as returned by optimx.

conv.code

the convergence code as returned by optimx.

log.likelihood

the value of the log-likelihood function at the found solution for the provided data.

See also