Fitting Linear Models with one Endogenous Regressor using Latent Instrumental Variables

Fits linear models with one endogenous regressor and no additional explanatory variables using the latent instrumental variable approach presented in Ebbes, P., Wedel, M., Böckenholt, U., and Steerneman, A. G. M. (2005). This is a statistical technique to address the endogeneity problem where no external instrumental variables are needed. The important assumption of the model is that the latent variables are discrete with at least two groups with different means and the structural error is normally distributed.

latentIV(
  formula,
  data,
  start.params = c(),
  optimx.args = list(),
  verbose = TRUE
)

Arguments

formula
A symbolic description of the model to be fitted. Of class "formula".

data
A data.frame containing the data of all parts specified in the formula parameter.

start.params
A named vector containing a set of parameters to use in the first optimization iteration. The names have to correspond exactly to the names of the components specified in the formula parameter. If not provided, a linear model is fitted to derive them.

optimx.args
A named list of arguments which are passed to optimx. This allows users to tweak optimization settings to their liking.

verbose
Show details about the running of the function.

Value

An object of classes rendo.latent.IV and rendo.base is returned which is a list and contains the following components:

formula

The formula given to specify the fitted model.

terms

The terms object used for model fitting.

model

The model.frame used for model fitting.

coefficients

A named vector of all coefficients resulting from model fitting.

names.main.coefs

a vector specifying which coefficients are from the model. For internal usage.

start.params

A named vector with the initial set of parameters used to optimize the log-likelihood function.

res.optimx

The result object returned by the function optimx after optimizing the log-likelihood function.

hessian

A named, symmetric matrix giving an estimate of the Hessian at the found solution.

m.delta.diag

A diagonal matrix needed when deriving the vcov to apply the delta method on theta5 which was transformed during the LL optimization.

fitted.values

Fitted values at the found optimal solution.

residuals

The residuals at the found optimal solution.

The function summary can be used to obtain and print a summary of the results. The generic accessor functions coefficients, fitted.values, residuals, vcov, confint, logLik, AIC, BIC, case.names, and nobs are available.
Details

Let's consider the model:
Yt=β0+αPt+εt Pt=π'Zt+νt

where \(t = 1,..,T\) indexes either time or cross-sectional units, Yt is the dependent variable, Pt is a k x 1 continuous, endogenous regressor, εt is a structural error term with mean zero and E(ε2)=σε2, \(\alpha\) and β0 are model parameters. Z;t is a l x 1 vector of instruments, and νt is a random error with mean zero and E(ν2)=σν2. The endogeneity problem arises from the correlation of \(P\) and εt through E(εν)=σεν

latentIV considers Zt' to be a latent, discrete, exogenous variable with an unknown number of groups \(m\) and \(\pi\) is a vector of group means. It is assumed that \(Z\) is independent of the error terms \(\epsilon\) and \(\nu\) and that it has at least two groups with different means. The structural and random errors are considered normally distributed with mean zero and variance-covariance matrix \(\Sigma\): Σ=(σε2, σ02,
σ02, σν2)

The identification of the model lies in the assumption of the non-normality of Pt, the discreteness of the unobserved instruments and the existence of at least two groups with different means.

The method has been implemented such that the latent variable has two groups. Ebbes et al.(2005) show in a Monte Carlo experiment that even if the true number of the categories of the instrument is larger than two, estimates are approximately consistent. Besides, overfitting in terms of the number of groups/categories reduces the degrees of freedom and leads to efficiency loss. For a model with additional explanatory variables a Bayesian approach is needed, since in a frequentist approach identification issues appear.

Identification of the parameters relies on the distributional assumptions of the latent instruments as well as that of the endogenous regressor Pt. Specifically, the endogenous regressor should have a non-normal distribution while the unobserved instruments, \(Z\), should be discrete and have at least two groups with different means Ebbes, Wedel, and Böckenholt (2009). A continuous distribution for the instruments leads to an unidentified model, while a normal distribution of the endogenous regressor gives rise to inefficient estimates.

Additional parameters used during model fitting and printed in summary are:

pi1

The instrumental variables \(Z\) are assumed to be divided into two groups. pi1 represents the estimated group mean of the first group.

pi2

The estimated group mean of the second group of the instrumental variables \(Z\).

theta5

The probability of being in the first group of the instruments.

theta6

The variance, σε2

theta7

The covariance, σεν

theta8

The variance, σν2

References

Ebbes, P., Wedel,M., Böckenholt, U., and Steerneman, A. G. M. (2005). 'Solving and Testing for Regressor-Error (in)Dependence When no Instrumental Variables are Available: With New Evidence for the Effect of Education on Income'. Quantitative Marketing and Economics, 3:365--392.

Ebbes P., Wedel M., Böckenholt U. (2009). “Frugal IV Alternatives to Identify the Parameter for an Endogenous Regressor.” Journal of Applied Econometrics, 24(3), 446–468.

See also

Examples

# \donttest{
data("dataLatentIV")

# function call without any initial parameter values
l  <- latentIV(y ~ P, data = dataLatentIV)

#> No start parameters were given. The linear model y ~ P is fitted to derive them.

#> The start parameters c((Intercept)=2.915, P=-0.852, pi1=0.589, pi2=2.001, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

summary(l)

#> 
#> Call:
#> latentIV(formula = y ~ P, data = dataLatentIV)
#> 
#> Coefficients:
#>             Estimate Std. Error z-score Pr(>|z|)    
#> (Intercept)  2.98632    0.03031   98.53   <2e-16 ***
#> P           -0.97539    0.03843  -25.38   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Further parameters estimated during model fitting:
#>     pi1     pi2  theta5  theta6  theta7  theta8 
#> 0.01744 2.21266 0.74306 1.01343 0.26144 1.09164 
#> (see help file for details)
#> 
#> Initial parameter values:
#> (Intercept)=2.9152 P=-0.8519 pi1=0.589 pi2=2.0011 theta5=0.5
#> theta6=1 theta7=0.5 theta8=1
#> 
#> The value of the log-likelihood function: 7873.344 
#> AIC: -15730.69 , BIC: -15684.1 
#> KKT1: FALSE  KKT2: TRUE Optimx Convergence Code: 0 

# function call with initial parameter values given by the user
l1 <- latentIV(y ~ P, start.params = c("(Intercept)"=2.5, P=-0.5),
               data = dataLatentIV)

#> The start parameters c((Intercept)=2.5, P=-0.5, pi1=0.589, pi2=2.001, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

summary(l1)

#> 
#> Call:
#> latentIV(formula = y ~ P, data = dataLatentIV, start.params = c(`(Intercept)` = 2.5, 
#>     P = -0.5))
#> 
#> Coefficients:
#>             Estimate Std. Error z-score Pr(>|z|)    
#> (Intercept)  2.97611    0.02949  100.91   <2e-16 ***
#> P           -0.94434    0.03540  -26.68   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Further parameters estimated during model fitting:
#>     pi1     pi2  theta5  theta6  theta7  theta8 
#> 0.01376 2.18676 0.72090 0.97086 0.18639 1.05935 
#> (see help file for details)
#> 
#> Initial parameter values:
#> (Intercept)=2.5 P=-0.5 pi1=0.589 pi2=2.0011 theta5=0.5 theta6=1
#> theta7=0.5 theta8=1
#> 
#> The value of the log-likelihood function: 7873.591 
#> AIC: -15731.18 , BIC: -15684.59 
#> KKT1: FALSE  KKT2: TRUE Optimx Convergence Code: 0 

# use own optimization settings (see optimx())
# set maximum number of iterations to 50'000
l2 <- latentIV(y ~ P, optimx.args = list(itnmax = 50000),
               data = dataLatentIV)

#> No start parameters were given. The linear model y ~ P is fitted to derive them.

#> The start parameters c((Intercept)=2.915, P=-0.852, pi1=0.589, pi2=2.001, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

# print detailed tracing information on progress
l3 <- latentIV(y ~ P, optimx.args = list(control = list(trace = 6)),
               data = dataLatentIV)

#> No start parameters were given. The linear model y ~ P is fitted to derive them.

#> The start parameters c((Intercept)=2.915, P=-0.852, pi1=0.589, pi2=2.001, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

#> fn is  fn 
#> Looking for method =  Nelder-Mead 
#> Methods to be used:[1] "Nelder-Mead"
#> Function has  8  arguments
#> Analytic gradient not made available.
#> Analytic Hessian not made available.
#> Scale check -- log parameter ratio= 0.7657037   log bounds ratio= NA 
#> optcfg:$fname
#> [1] "fn"
#> 
#> $npar
#> [1] 8
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 6
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] TRUE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (params, m.data.mvnorm, use.intercept, name.intercept, 
#>     name.endo.param) 
#> {
#>     if (use.intercept) {
#>         b00 <- params[name.intercept]
#>         a1 <- params[name.endo.param]
#>     }
#>     else {
#>         b00 <- 0
#>         a1 <- params[name.endo.param]
#>     }
#>     pt <- exp(params["theta5"])
#>     pt <- pt/(1 + pt)
#>     m.sigma <- matrix(c(params["theta6"], params["theta7"], params["theta7"], 
#>         params["theta8"]), byrow = TRUE, ncol = 2, nrow = 2)
#>     s2e <- m.sigma[1, 1]
#>     s2v <- m.sigma[2, 2]
#>     sev <- m.sigma[1, 2]
#>     varcov <- matrix(0, 2, 2)
#>     varcov[1, 1] <- a1 * a1 * s2v + 2 * a1 * sev + s2e
#>     varcov[2, 1] <- a1 * s2v + sev
#>     varcov[1, 2] <- varcov[2, 1]
#>     varcov[2, 2] <- s2v
#>     pi1 <- params["pi1"]
#>     mu1 <- matrix(data = c(b00 + a1 * pi1, pi1), nrow = 2, ncol = 1)
#>     pi2 <- params["pi2"]
#>     mu2 <- matrix(c(b00 + a1 * pi2, pi2), nrow = 2, ncol = 1)
#>     log.pdf1 <- dmvnorm(m.data.mvnorm, mean = mu1, sigma = varcov, 
#>         log = TRUE)
#>     log.pdf2 <- dmvnorm(m.data.mvnorm, mean = mu2, sigma = varcov, 
#>         log = TRUE)
#>     max.AB <- pmax(log(pt) + log.pdf1, log(1 - pt) + log.pdf2)
#>     logLL.lse <- sum(max.AB + log(pt * exp(log.pdf1 - max.AB) + 
#>         (1 - pt) * exp(log.pdf2 - max.AB)))
#>     return(-1 * logLL.lse)
#> }
#> <bytecode: 0x7fc0fa718390>
#> <environment: namespace:REndo>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "Nelder-Mead"
#> 
#> Method:  Nelder-Mead 
#>   Nelder-Mead direct search function minimizer
#> function value for initial parameters = 8742.932040
#>   Scaled convergence tolerance is 0.00013028
#> Stepsize computed as 0.291524
#> BUILD              9 11102.056553 8424.840299
#> EXTENSION         11 9624.462289 8325.753007
#> REFLECTION        13 9272.764194 8305.555394
#> EXTENSION         15 8742.932040 7944.347065
#> LO-REDUCTION      17 8703.872300 7944.347065
#> LO-REDUCTION      19 8690.122323 7944.347065
#> LO-REDUCTION      21 8680.249246 7944.347065
#> LO-REDUCTION      23 8592.719293 7944.347065
#> LO-REDUCTION      25 8424.840299 7944.347065
#> LO-REDUCTION      27 8325.753007 7944.347065
#> LO-REDUCTION      29 8305.555394 7944.347065
#> LO-REDUCTION      31 8130.497227 7944.347065
#> LO-REDUCTION      33 8114.012489 7944.347065
#> LO-REDUCTION      35 8113.150479 7944.347065
#> REFLECTION        37 8081.291967 7937.018590
#> LO-REDUCTION      39 8077.875831 7937.018590
#> LO-REDUCTION      41 8051.362993 7934.686437
#> LO-REDUCTION      43 8011.993137 7924.134452
#> REFLECTION        45 7957.286375 7906.855907
#> HI-REDUCTION      47 7951.085472 7906.855907
#> LO-REDUCTION      49 7946.239355 7906.855907
#> LO-REDUCTION      51 7944.347065 7906.855907
#> LO-REDUCTION      53 7937.652502 7906.855907
#> LO-REDUCTION      55 7937.018590 7906.855907
#> REFLECTION        57 7934.686437 7905.372810
#> REFLECTION        59 7927.052505 7902.136181
#> LO-REDUCTION      61 7924.134452 7902.136181
#> REFLECTION        63 7921.069576 7893.709734
#> LO-REDUCTION      65 7917.698438 7893.709734
#> REFLECTION        67 7916.557036 7892.335377
#> REFLECTION        69 7910.240349 7891.295754
#> HI-REDUCTION      71 7908.458846 7891.295754
#> LO-REDUCTION      73 7906.855907 7890.080494
#> LO-REDUCTION      75 7905.372810 7890.080494
#> LO-REDUCTION      77 7902.136181 7890.080494
#> LO-REDUCTION      79 7898.525855 7890.080494
#> LO-REDUCTION      81 7897.809519 7890.080494
#> EXTENSION         83 7894.215378 7884.644817
#> HI-REDUCTION      85 7893.709734 7884.644817
#> LO-REDUCTION      87 7893.447130 7884.644817
#> LO-REDUCTION      89 7892.335377 7884.644817
#> HI-REDUCTION      91 7891.295754 7884.644817
#> LO-REDUCTION      93 7890.923583 7884.644817
#> REFLECTION        95 7890.221386 7884.554251
#> LO-REDUCTION      97 7890.080494 7884.554251
#> LO-REDUCTION      99 7889.589323 7884.554251
#> REFLECTION       101 7888.284750 7883.793389
#> LO-REDUCTION     103 7886.969431 7883.793389
#> REFLECTION       105 7886.766207 7883.060300
#> HI-REDUCTION     107 7886.700101 7883.060300
#> REFLECTION       109 7885.826209 7883.036358
#> REFLECTION       111 7884.841099 7882.527521
#> REFLECTION       113 7884.675015 7882.360551
#> LO-REDUCTION     115 7884.644817 7882.360551
#> REFLECTION       117 7884.554251 7881.962927
#> HI-REDUCTION     119 7884.461336 7881.962927
#> EXTENSION        121 7883.793389 7880.668711
#> LO-REDUCTION     123 7883.060300 7880.668711
#> LO-REDUCTION     125 7883.036358 7880.668711
#> LO-REDUCTION     127 7883.013367 7880.668711
#> LO-REDUCTION     129 7882.890033 7880.668711
#> LO-REDUCTION     131 7882.527521 7880.668711
#> LO-REDUCTION     133 7882.360551 7880.668711
#> LO-REDUCTION     135 7882.044811 7880.668711
#> LO-REDUCTION     137 7881.962927 7880.668711
#> LO-REDUCTION     139 7881.540015 7880.668711
#> REFLECTION       141 7881.364098 7880.506212
#> LO-REDUCTION     143 7881.209701 7880.506212
#> LO-REDUCTION     145 7881.043959 7880.506212
#> REFLECTION       147 7881.015615 7880.461745
#> LO-REDUCTION     149 7880.945479 7880.461745
#> LO-REDUCTION     151 7880.860779 7880.461745
#> LO-REDUCTION     153 7880.765993 7880.461745
#> LO-REDUCTION     155 7880.727698 7880.461745
#> EXTENSION        157 7880.685065 7880.292296
#> LO-REDUCTION     159 7880.668711 7880.292296
#> LO-REDUCTION     161 7880.628690 7880.292296
#> LO-REDUCTION     163 7880.569007 7880.292296
#> LO-REDUCTION     165 7880.565565 7880.292296
#> EXTENSION        167 7880.557113 7880.246277
#> EXTENSION        169 7880.506212 7880.152925
#> EXTENSION        171 7880.463658 7880.117354
#> EXTENSION        173 7880.461745 7880.012803
#> LO-REDUCTION     175 7880.355627 7880.012803
#> EXTENSION        177 7880.334173 7879.829468
#> LO-REDUCTION     179 7880.332691 7879.829468
#> EXTENSION        181 7880.292296 7879.660900
#> LO-REDUCTION     183 7880.246277 7879.660900
#> LO-REDUCTION     185 7880.152925 7879.660900
#> REFLECTION       187 7880.117354 7879.617674
#> HI-REDUCTION     189 7880.117139 7879.617674
#> EXTENSION        191 7880.012803 7879.268060
#> LO-REDUCTION     193 7879.914322 7879.268060
#> LO-REDUCTION     195 7879.829468 7879.268060
#> LO-REDUCTION     197 7879.799598 7879.268060
#> LO-REDUCTION     199 7879.783529 7879.268060
#> LO-REDUCTION     201 7879.753101 7879.268060
#> LO-REDUCTION     203 7879.660900 7879.268060
#> LO-REDUCTION     205 7879.617674 7879.268060
#> LO-REDUCTION     207 7879.585480 7879.268060
#> LO-REDUCTION     209 7879.561489 7879.268060
#> EXTENSION        211 7879.438207 7879.168968
#> EXTENSION        213 7879.435870 7879.132444
#> LO-REDUCTION     215 7879.428865 7879.132444
#> EXTENSION        217 7879.385164 7878.968881
#> LO-REDUCTION     219 7879.340957 7878.968881
#> LO-REDUCTION     221 7879.305036 7878.968881
#> EXTENSION        223 7879.281966 7878.760648
#> LO-REDUCTION     225 7879.268060 7878.760648
#> LO-REDUCTION     227 7879.195136 7878.760648
#> LO-REDUCTION     229 7879.168968 7878.760648
#> LO-REDUCTION     231 7879.132444 7878.760648
#> LO-REDUCTION     233 7879.059316 7878.760648
#> LO-REDUCTION     235 7878.975654 7878.760648
#> EXTENSION        237 7878.971720 7878.655217
#> LO-REDUCTION     239 7878.968881 7878.655217
#> EXTENSION        241 7878.903387 7878.616791
#> EXTENSION        243 7878.844085 7878.494173
#> LO-REDUCTION     245 7878.827880 7878.494173
#> LO-REDUCTION     247 7878.808252 7878.494173
#> REFLECTION       249 7878.800476 7878.442731
#> REFLECTION       251 7878.760648 7878.408916
#> LO-REDUCTION     253 7878.715302 7878.408916
#> EXTENSION        255 7878.655217 7878.025805
#> LO-REDUCTION     257 7878.616791 7878.025805
#> LO-REDUCTION     259 7878.571245 7878.025805
#> LO-REDUCTION     261 7878.509654 7878.025805
#> LO-REDUCTION     263 7878.494173 7878.025805
#> LO-REDUCTION     265 7878.442731 7878.025805
#> LO-REDUCTION     267 7878.438463 7878.025805
#> EXTENSION        269 7878.408916 7877.760970
#> LO-REDUCTION     271 7878.342853 7877.760970
#> LO-REDUCTION     273 7878.329483 7877.760970
#> EXTENSION        275 7878.266674 7877.596564
#> LO-REDUCTION     277 7878.144760 7877.596564
#> LO-REDUCTION     279 7878.132136 7877.596564
#> EXTENSION        281 7878.041083 7877.397780
#> LO-REDUCTION     283 7878.025805 7877.397780
#> EXTENSION        285 7877.983724 7877.329001
#> EXTENSION        287 7877.918973 7877.238097
#> REFLECTION       289 7877.760970 7877.212135
#> LO-REDUCTION     291 7877.670399 7877.212135
#> LO-REDUCTION     293 7877.629038 7877.212135
#> LO-REDUCTION     295 7877.596564 7877.212135
#> LO-REDUCTION     297 7877.507114 7877.212135
#> REFLECTION       299 7877.397780 7877.196937
#> HI-REDUCTION     301 7877.329001 7877.196937
#> REFLECTION       303 7877.295556 7877.181607
#> EXTENSION        305 7877.262602 7876.954473
#> LO-REDUCTION     307 7877.258082 7876.954473
#> LO-REDUCTION     309 7877.247359 7876.954473
#> LO-REDUCTION     311 7877.238097 7876.954473
#> LO-REDUCTION     313 7877.230412 7876.954473
#> LO-REDUCTION     315 7877.212135 7876.954473
#> LO-REDUCTION     317 7877.196937 7876.954473
#> LO-REDUCTION     319 7877.181607 7876.954473
#> EXTENSION        321 7877.110460 7876.856400
#> EXTENSION        323 7877.089184 7876.739058
#> LO-REDUCTION     325 7877.082864 7876.739058
#> EXTENSION        327 7877.053283 7876.549872
#> LO-REDUCTION     329 7876.986577 7876.549872
#> LO-REDUCTION     331 7876.986009 7876.549872
#> LO-REDUCTION     333 7876.980204 7876.549872
#> LO-REDUCTION     335 7876.954473 7876.549872
#> EXTENSION        337 7876.856400 7876.237010
#> LO-REDUCTION     339 7876.787029 7876.237010
#> LO-REDUCTION     341 7876.739058 7876.237010
#> LO-REDUCTION     343 7876.687658 7876.237010
#> LO-REDUCTION     345 7876.671107 7876.237010
#> EXTENSION        347 7876.633511 7876.153356
#> EXTENSION        349 7876.551829 7876.031788
#> EXTENSION        351 7876.549872 7875.881384
#> LO-REDUCTION     353 7876.513139 7875.881384
#> EXTENSION        355 7876.413540 7875.562368
#> LO-REDUCTION     357 7876.311302 7875.562368
#> LO-REDUCTION     359 7876.275977 7875.562368
#> LO-REDUCTION     361 7876.237010 7875.562368
#> LO-REDUCTION     363 7876.153356 7875.562368
#> LO-REDUCTION     365 7876.031788 7875.562368
#> LO-REDUCTION     367 7875.992399 7875.562368
#> EXTENSION        369 7875.889476 7875.437998
#> EXTENSION        371 7875.881384 7875.214806
#> LO-REDUCTION     373 7875.839502 7875.214806
#> LO-REDUCTION     375 7875.799629 7875.214806
#> EXTENSION        377 7875.779035 7874.979370
#> LO-REDUCTION     379 7875.673535 7874.979370
#> LO-REDUCTION     381 7875.657270 7874.979370
#> EXTENSION        383 7875.562368 7874.951503
#> LO-REDUCTION     385 7875.438713 7874.951503
#> EXTENSION        387 7875.437998 7874.595900
#> EXTENSION        389 7875.304646 7874.444334
#> LO-REDUCTION     391 7875.246611 7874.444334
#> EXTENSION        393 7875.214806 7874.127223
#> LO-REDUCTION     395 7875.046483 7874.127223
#> LO-REDUCTION     397 7875.012419 7874.127223
#> LO-REDUCTION     399 7874.979370 7874.127223
#> LO-REDUCTION     401 7874.951503 7874.127223
#> LO-REDUCTION     403 7874.595900 7874.127223
#> LO-REDUCTION     405 7874.566430 7874.127223
#> EXTENSION        407 7874.495020 7874.101348
#> EXTENSION        409 7874.492839 7873.967425
#> LO-REDUCTION     411 7874.444334 7873.967425
#> REFLECTION       413 7874.343326 7873.921079
#> REFLECTION       415 7874.310502 7873.850982
#> LO-REDUCTION     417 7874.219801 7873.850982
#> LO-REDUCTION     419 7874.138834 7873.850982
#> LO-REDUCTION     421 7874.127223 7873.850982
#> LO-REDUCTION     423 7874.101348 7873.850982
#> EXTENSION        425 7874.081124 7873.783880
#> LO-REDUCTION     427 7873.983422 7873.783880
#> LO-REDUCTION     429 7873.982819 7873.783880
#> LO-REDUCTION     431 7873.977598 7873.783880
#> EXTENSION        433 7873.967425 7873.747915
#> LO-REDUCTION     435 7873.964820 7873.747915
#> LO-REDUCTION     437 7873.921079 7873.742302
#> LO-REDUCTION     439 7873.850982 7873.742302
#> LO-REDUCTION     441 7873.843559 7873.742302
#> EXTENSION        443 7873.838925 7873.670492
#> LO-REDUCTION     445 7873.826824 7873.670492
#> LO-REDUCTION     447 7873.783880 7873.670492
#> LO-REDUCTION     449 7873.768364 7873.670492
#> LO-REDUCTION     451 7873.752220 7873.670492
#> HI-REDUCTION     453 7873.747915 7873.670492
#> LO-REDUCTION     455 7873.743085 7873.670492
#> LO-REDUCTION     457 7873.742302 7873.670492
#> LO-REDUCTION     459 7873.727311 7873.670492
#> REFLECTION       461 7873.706839 7873.663163
#> LO-REDUCTION     463 7873.694156 7873.663163
#> LO-REDUCTION     465 7873.693556 7873.663163
#> LO-REDUCTION     467 7873.693149 7873.663163
#> LO-REDUCTION     469 7873.680706 7873.663163
#> REFLECTION       471 7873.679782 7873.658683
#> LO-REDUCTION     473 7873.677118 7873.658683
#> HI-REDUCTION     475 7873.673279 7873.658683
#> EXTENSION        477 7873.670492 7873.649404
#> LO-REDUCTION     479 7873.668406 7873.649404
#> EXTENSION        481 7873.667615 7873.645688
#> LO-REDUCTION     483 7873.665794 7873.645688
#> LO-REDUCTION     485 7873.665673 7873.645688
#> LO-REDUCTION     487 7873.663163 7873.645688
#> REFLECTION       489 7873.661091 7873.642268
#> EXTENSION        491 7873.658683 7873.625179
#> LO-REDUCTION     493 7873.657071 7873.625179
#> LO-REDUCTION     495 7873.652898 7873.625179
#> LO-REDUCTION     497 7873.649954 7873.625179
#> LO-REDUCTION     499 7873.649404 7873.625179
#> EXTENSION        501 7873.645933 7873.617972
#> EXTENSION        503 7873.645688 7873.609255
#> LO-REDUCTION     505 7873.642268 7873.609255
#> LO-REDUCTION     507 7873.642057 7873.609255
#> EXTENSION        509 7873.635016 7873.582029
#> LO-REDUCTION     511 7873.627337 7873.582029
#> LO-REDUCTION     513 7873.625601 7873.582029
#> LO-REDUCTION     515 7873.625179 7873.582029
#> EXTENSION        517 7873.618683 7873.574109
#> LO-REDUCTION     519 7873.617972 7873.574109
#> LO-REDUCTION     521 7873.615075 7873.574109
#> LO-REDUCTION     523 7873.609255 7873.574109
#> REFLECTION       525 7873.599968 7873.573347
#> EXTENSION        527 7873.599411 7873.544108
#> LO-REDUCTION     529 7873.587268 7873.544108
#> LO-REDUCTION     531 7873.585021 7873.544108
#> LO-REDUCTION     533 7873.582029 7873.544108
#> LO-REDUCTION     535 7873.580517 7873.544108
#> LO-REDUCTION     537 7873.575567 7873.544108
#> LO-REDUCTION     539 7873.574109 7873.544108
#> LO-REDUCTION     541 7873.573347 7873.544108
#> EXTENSION        543 7873.557222 7873.521072
#> LO-REDUCTION     545 7873.556680 7873.521072
#> LO-REDUCTION     547 7873.554481 7873.521072
#> LO-REDUCTION     549 7873.552024 7873.521072
#> LO-REDUCTION     551 7873.551309 7873.521072
#> EXTENSION        553 7873.549423 7873.502281
#> LO-REDUCTION     555 7873.544967 7873.502281
#> LO-REDUCTION     557 7873.544108 7873.502281
#> LO-REDUCTION     559 7873.533628 7873.502281
#> LO-REDUCTION     561 7873.529218 7873.502281
#> REFLECTION       563 7873.528686 7873.501992
#> REFLECTION       565 7873.526832 7873.501979
#> REFLECTION       567 7873.521072 7873.496847
#> LO-REDUCTION     569 7873.513217 7873.496847
#> REFLECTION       571 7873.510430 7873.495121
#> EXTENSION        573 7873.506431 7873.490333
#> LO-REDUCTION     575 7873.502946 7873.490333
#> REFLECTION       577 7873.502281 7873.488898
#> LO-REDUCTION     579 7873.501992 7873.488898
#> LO-REDUCTION     581 7873.501979 7873.488898
#> LO-REDUCTION     583 7873.497228 7873.488898
#> HI-REDUCTION     585 7873.496847 7873.488898
#> LO-REDUCTION     587 7873.495121 7873.488898
#> HI-REDUCTION     589 7873.494892 7873.488898
#> REFLECTION       591 7873.493812 7873.488551
#> EXTENSION        593 7873.492191 7873.484993
#> LO-REDUCTION     595 7873.491183 7873.484993
#> LO-REDUCTION     597 7873.490614 7873.484993
#> LO-REDUCTION     599 7873.490333 7873.484993
#> LO-REDUCTION     601 7873.489726 7873.484993
#> LO-REDUCTION     603 7873.488979 7873.484993
#> REFLECTION       605 7873.488898 7873.484879
#> LO-REDUCTION     607 7873.488551 7873.484879
#> LO-REDUCTION     609 7873.487332 7873.484739
#> REFLECTION       611 7873.487115 7873.483926
#> LO-REDUCTION     613 7873.487042 7873.483926
#> LO-REDUCTION     615 7873.486005 7873.483926
#> EXTENSION        617 7873.485945 7873.481990
#> LO-REDUCTION     619 7873.485063 7873.481990
#> LO-REDUCTION     621 7873.484993 7873.481990
#> LO-REDUCTION     623 7873.484879 7873.481990
#> LO-REDUCTION     625 7873.484739 7873.481990
#> LO-REDUCTION     627 7873.484430 7873.481990
#> EXTENSION        629 7873.483966 7873.480264
#> LO-REDUCTION     631 7873.483926 7873.480264
#> LO-REDUCTION     633 7873.483886 7873.480264
#> LO-REDUCTION     635 7873.483686 7873.480264
#> EXTENSION        637 7873.483379 7873.478829
#> LO-REDUCTION     639 7873.483198 7873.478829
#> LO-REDUCTION     641 7873.482537 7873.478829
#> LO-REDUCTION     643 7873.481990 7873.478829
#> EXTENSION        645 7873.481699 7873.477523
#> LO-REDUCTION     647 7873.481560 7873.477523
#> EXTENSION        649 7873.481291 7873.474958
#> LO-REDUCTION     651 7873.480264 7873.474958
#> LO-REDUCTION     653 7873.480125 7873.474958
#> LO-REDUCTION     655 7873.479825 7873.474958
#> EXTENSION        657 7873.478932 7873.471649
#> LO-REDUCTION     659 7873.478829 7873.471649
#> LO-REDUCTION     661 7873.477619 7873.471649
#> LO-REDUCTION     663 7873.477523 7873.471649
#> LO-REDUCTION     665 7873.476761 7873.471649
#> LO-REDUCTION     667 7873.476718 7873.471649
#> EXTENSION        669 7873.476336 7873.470920
#> EXTENSION        671 7873.474958 7873.467269
#> LO-REDUCTION     673 7873.474719 7873.467269
#> LO-REDUCTION     675 7873.474430 7873.467269
#> EXTENSION        677 7873.473312 7873.464401
#> LO-REDUCTION     679 7873.473288 7873.464401
#> LO-REDUCTION     681 7873.472276 7873.464401
#> LO-REDUCTION     683 7873.471649 7873.464401
#> LO-REDUCTION     685 7873.470920 7873.464401
#> EXTENSION        687 7873.470326 7873.460417
#> LO-REDUCTION     689 7873.469271 7873.460417
#> LO-REDUCTION     691 7873.467269 7873.460417
#> LO-REDUCTION     693 7873.467147 7873.460417
#> EXTENSION        695 7873.466899 7873.456121
#> LO-REDUCTION     697 7873.466173 7873.456121
#> LO-REDUCTION     699 7873.465248 7873.456121
#> EXTENSION        701 7873.464401 7873.454121
#> LO-REDUCTION     703 7873.462947 7873.454121
#> EXTENSION        705 7873.461762 7873.451189
#> LO-REDUCTION     707 7873.461244 7873.451189
#> EXTENSION        709 7873.460417 7873.448989
#> EXTENSION        711 7873.458872 7873.443598
#> LO-REDUCTION     713 7873.457406 7873.443598
#> LO-REDUCTION     715 7873.456121 7873.443598
#> LO-REDUCTION     717 7873.454768 7873.443598
#> LO-REDUCTION     719 7873.454121 7873.443598
#> LO-REDUCTION     721 7873.452064 7873.443598
#> LO-REDUCTION     723 7873.451189 7873.443598
#> LO-REDUCTION     725 7873.449800 7873.443598
#> LO-REDUCTION     727 7873.449744 7873.443598
#> LO-REDUCTION     729 7873.448989 7873.443598
#> REFLECTION       731 7873.446894 7873.442969
#> EXTENSION        733 7873.446574 7873.442182
#> EXTENSION        735 7873.446001 7873.438511
#> LO-REDUCTION     737 7873.445895 7873.438511
#> EXTENSION        739 7873.445438 7873.435462
#> LO-REDUCTION     741 7873.444508 7873.435462
#> LO-REDUCTION     743 7873.443663 7873.435462
#> LO-REDUCTION     745 7873.443598 7873.435462
#> EXTENSION        747 7873.442969 7873.434681
#> LO-REDUCTION     749 7873.442182 7873.434681
#> EXTENSION        751 7873.440886 7873.427479
#> LO-REDUCTION     753 7873.438784 7873.427479
#> LO-REDUCTION     755 7873.438511 7873.427479
#> LO-REDUCTION     757 7873.437999 7873.427479
#> LO-REDUCTION     759 7873.436864 7873.427479
#> EXTENSION        761 7873.436042 7873.427155
#> EXTENSION        763 7873.435462 7873.423607
#> LO-REDUCTION     765 7873.434681 7873.423607
#> EXTENSION        767 7873.431663 7873.415235
#> LO-REDUCTION     769 7873.430635 7873.415235
#> LO-REDUCTION     771 7873.429266 7873.415235
#> LO-REDUCTION     773 7873.428377 7873.415235
#> LO-REDUCTION     775 7873.427479 7873.415235
#> LO-REDUCTION     777 7873.427155 7873.415235
#> LO-REDUCTION     779 7873.426117 7873.415235
#> EXTENSION        781 7873.423678 7873.413749
#> EXTENSION        783 7873.423607 7873.410755
#> EXTENSION        785 7873.421180 7873.409932
#> REFLECTION       787 7873.420377 7873.408827
#> LO-REDUCTION     789 7873.419484 7873.408827
#> LO-REDUCTION     791 7873.419444 7873.408827
#> LO-REDUCTION     793 7873.416098 7873.408827
#> REFLECTION       795 7873.415235 7873.408700
#> HI-REDUCTION     797 7873.413749 7873.408700
#> LO-REDUCTION     799 7873.410755 7873.408700
#> LO-REDUCTION     801 7873.410382 7873.408700
#> HI-REDUCTION     803 7873.409996 7873.408700
#> LO-REDUCTION     805 7873.409932 7873.408618
#> REFLECTION       807 7873.409775 7873.408476
#> REFLECTION       809 7873.409367 7873.408224
#> EXTENSION        811 7873.409104 7873.407191
#> LO-REDUCTION     813 7873.409007 7873.407191
#> LO-REDUCTION     815 7873.409002 7873.407191
#> REFLECTION       817 7873.408827 7873.407100
#> LO-REDUCTION     819 7873.408700 7873.407100
#> LO-REDUCTION     821 7873.408618 7873.407100
#> EXTENSION        823 7873.408476 7873.404994
#> LO-REDUCTION     825 7873.408224 7873.404994
#> LO-REDUCTION     827 7873.407767 7873.404994
#> LO-REDUCTION     829 7873.407584 7873.404994
#> LO-REDUCTION     831 7873.407297 7873.404994
#> LO-REDUCTION     833 7873.407191 7873.404994
#> LO-REDUCTION     835 7873.407147 7873.404994
#> EXTENSION        837 7873.407100 7873.403559
#> EXTENSION        839 7873.406847 7873.402593
#> LO-REDUCTION     841 7873.405530 7873.402593
#> LO-REDUCTION     843 7873.405389 7873.402593
#> LO-REDUCTION     845 7873.405383 7873.402593
#> EXTENSION        847 7873.405068 7873.400300
#> LO-REDUCTION     849 7873.405001 7873.400300
#> LO-REDUCTION     851 7873.404994 7873.400300
#> REFLECTION       853 7873.403559 7873.399970
#> LO-REDUCTION     855 7873.403085 7873.399970
#> LO-REDUCTION     857 7873.402983 7873.399970
#> EXTENSION        859 7873.402648 7873.398938
#> EXTENSION        861 7873.402593 7873.397505
#> LO-REDUCTION     863 7873.401110 7873.397505
#> LO-REDUCTION     865 7873.400933 7873.397505
#> LO-REDUCTION     867 7873.400567 7873.397505
#> REFLECTION       869 7873.400371 7873.397392
#> REFLECTION       871 7873.400300 7873.397190
#> LO-REDUCTION     873 7873.399970 7873.397190
#> LO-REDUCTION     875 7873.398938 7873.397190
#> LO-REDUCTION     877 7873.398449 7873.397190
#> EXTENSION        879 7873.398253 7873.396296
#> LO-REDUCTION     881 7873.397865 7873.396296
#> LO-REDUCTION     883 7873.397552 7873.396296
#> LO-REDUCTION     885 7873.397505 7873.396296
#> HI-REDUCTION     887 7873.397469 7873.396296
#> REFLECTION       889 7873.397392 7873.396167
#> LO-REDUCTION     891 7873.397225 7873.396167
#> EXTENSION        893 7873.397190 7873.395257
#> EXTENSION        895 7873.396851 7873.394744
#> LO-REDUCTION     897 7873.396699 7873.394744
#> EXTENSION        899 7873.396469 7873.393910
#> LO-REDUCTION     901 7873.396373 7873.393910
#> LO-REDUCTION     903 7873.396296 7873.393910
#> LO-REDUCTION     905 7873.396293 7873.393910
#> EXTENSION        907 7873.396167 7873.393169
#> LO-REDUCTION     909 7873.395459 7873.393169
#> EXTENSION        911 7873.395257 7873.391089
#> LO-REDUCTION     913 7873.394837 7873.391089
#> LO-REDUCTION     915 7873.394760 7873.391089
#> LO-REDUCTION     917 7873.394744 7873.391089
#> LO-REDUCTION     919 7873.394281 7873.391089
#> LO-REDUCTION     921 7873.393910 7873.391089
#> LO-REDUCTION     923 7873.393608 7873.391089
#> LO-REDUCTION     925 7873.393405 7873.391089
#> EXTENSION        927 7873.393169 7873.390186
#> LO-REDUCTION     929 7873.392917 7873.390186
#> EXTENSION        931 7873.392244 7873.389692
#> LO-REDUCTION     933 7873.392234 7873.389692
#> EXTENSION        935 7873.392018 7873.387249
#> LO-REDUCTION     937 7873.391551 7873.387249
#> LO-REDUCTION     939 7873.391287 7873.387249
#> EXTENSION        941 7873.391089 7873.386046
#> LO-REDUCTION     943 7873.390381 7873.386046
#> EXTENSION        945 7873.390186 7873.384738
#> LO-REDUCTION     947 7873.389764 7873.384738
#> LO-REDUCTION     949 7873.389692 7873.384738
#> EXTENSION        951 7873.387794 7873.382773
#> LO-REDUCTION     953 7873.387778 7873.382773
#> LO-REDUCTION     955 7873.387267 7873.382773
#> EXTENSION        957 7873.387249 7873.380669
#> LO-REDUCTION     959 7873.386288 7873.380669
#> LO-REDUCTION     961 7873.386046 7873.380669
#> LO-REDUCTION     963 7873.384925 7873.380669
#> EXTENSION        965 7873.384738 7873.379951
#> LO-REDUCTION     967 7873.384083 7873.379951
#> EXTENSION        969 7873.383650 7873.377437
#> LO-REDUCTION     971 7873.383169 7873.377437
#> LO-REDUCTION     973 7873.382773 7873.377437
#> LO-REDUCTION     975 7873.382607 7873.377437
#> EXTENSION        977 7873.381259 7873.374642
#> LO-REDUCTION     979 7873.380669 7873.374642
#> HI-REDUCTION     981 7873.380010 7873.374642
#> LO-REDUCTION     983 7873.379951 7873.374642
#> LO-REDUCTION     985 7873.378990 7873.374642
#> LO-REDUCTION     987 7873.378600 7873.374642
#> LO-REDUCTION     989 7873.378321 7873.374642
#> LO-REDUCTION     991 7873.377814 7873.374642
#> LO-REDUCTION     993 7873.377736 7873.374642
#> EXTENSION        995 7873.377437 7873.371732
#> LO-REDUCTION     997 7873.376019 7873.371732
#> LO-REDUCTION     999 7873.375561 7873.371732
#> LO-REDUCTION    1001 7873.375208 7873.371732
#> LO-REDUCTION    1003 7873.374979 7873.371732
#> EXTENSION       1005 7873.374813 7873.371066
#> LO-REDUCTION    1007 7873.374664 7873.371066
#> LO-REDUCTION    1009 7873.374642 7873.371066
#> REFLECTION      1011 7873.373421 7873.370501
#> EXTENSION       1013 7873.372471 7873.368264
#> LO-REDUCTION    1015 7873.372039 7873.368264
#> LO-REDUCTION    1017 7873.372032 7873.368264
#> LO-REDUCTION    1019 7873.371732 7873.368264
#> LO-REDUCTION    1021 7873.371695 7873.368264
#> EXTENSION       1023 7873.371635 7873.367673
#> LO-REDUCTION    1025 7873.371066 7873.367673
#> LO-REDUCTION    1027 7873.370501 7873.367673
#> LO-REDUCTION    1029 7873.370031 7873.367673
#> EXTENSION       1031 7873.369563 7873.367261
#> EXTENSION       1033 7873.369368 7873.365073
#> LO-REDUCTION    1035 7873.369179 7873.365073
#> LO-REDUCTION    1037 7873.368427 7873.365073
#> EXTENSION       1039 7873.368264 7873.363784
#> LO-REDUCTION    1041 7873.368264 7873.363784
#> LO-REDUCTION    1043 7873.367830 7873.363784
#> LO-REDUCTION    1045 7873.367673 7873.363784
#> LO-REDUCTION    1047 7873.367261 7873.363784
#> EXTENSION       1049 7873.366801 7873.362781
#> EXTENSION       1051 7873.365493 7873.360155
#> LO-REDUCTION    1053 7873.365465 7873.360155
#> LO-REDUCTION    1055 7873.365415 7873.360155
#> LO-REDUCTION    1057 7873.365073 7873.360155
#> EXTENSION       1059 7873.364091 7873.359633
#> REFLECTION      1061 7873.363786 7873.359459
#> LO-REDUCTION    1063 7873.363784 7873.359459
#> EXTENSION       1065 7873.362781 7873.357881
#> LO-REDUCTION    1067 7873.361532 7873.357881
#> LO-REDUCTION    1069 7873.361154 7873.357881
#> LO-REDUCTION    1071 7873.360600 7873.357881
#> LO-REDUCTION    1073 7873.360155 7873.357881
#> LO-REDUCTION    1075 7873.359769 7873.357881
#> LO-REDUCTION    1077 7873.359633 7873.357881
#> REFLECTION      1079 7873.359459 7873.357849
#> REFLECTION      1081 7873.358683 7873.357736
#> EXTENSION       1083 7873.358268 7873.356832
#> LO-REDUCTION    1085 7873.358217 7873.356832
#> LO-REDUCTION    1087 7873.358177 7873.356832
#> REFLECTION      1089 7873.358082 7873.356818
#> EXTENSION       1091 7873.358023 7873.356123
#> HI-REDUCTION    1093 7873.357881 7873.356123
#> LO-REDUCTION    1095 7873.357849 7873.356123
#> LO-REDUCTION    1097 7873.357736 7873.356123
#> REFLECTION      1099 7873.357051 7873.356002
#> EXTENSION       1101 7873.357050 7873.355468
#> EXTENSION       1103 7873.356832 7873.354776
#> LO-REDUCTION    1105 7873.356818 7873.354776
#> LO-REDUCTION    1107 7873.356791 7873.354776
#> LO-REDUCTION    1109 7873.356509 7873.354776
#> REFLECTION      1111 7873.356382 7873.354713
#> LO-REDUCTION    1113 7873.356123 7873.354713
#> REFLECTION      1115 7873.356002 7873.354493
#> REFLECTION      1117 7873.355518 7873.354353
#> EXTENSION       1119 7873.355468 7873.353795
#> LO-REDUCTION    1121 7873.355237 7873.353795
#> LO-REDUCTION    1123 7873.355096 7873.353795
#> LO-REDUCTION    1125 7873.354916 7873.353795
#> EXTENSION       1127 7873.354776 7873.352967
#> LO-REDUCTION    1129 7873.354713 7873.352967
#> LO-REDUCTION    1131 7873.354493 7873.352967
#> EXTENSION       1133 7873.354353 7873.351907
#> LO-REDUCTION    1135 7873.354045 7873.351907
#> LO-REDUCTION    1137 7873.353875 7873.351907
#> LO-REDUCTION    1139 7873.353873 7873.351907
#> LO-REDUCTION    1141 7873.353795 7873.351907
#> EXTENSION       1143 7873.353153 7873.351113
#> LO-REDUCTION    1145 7873.353064 7873.351113
#> LO-REDUCTION    1147 7873.352967 7873.351113
#> REFLECTION      1149 7873.352753 7873.351074
#> EXTENSION       1151 7873.352681 7873.349926
#> LO-REDUCTION    1153 7873.352241 7873.349926
#> LO-REDUCTION    1155 7873.352143 7873.349926
#> LO-REDUCTION    1157 7873.351907 7873.349926
#> LO-REDUCTION    1159 7873.351181 7873.349926
#> LO-REDUCTION    1161 7873.351162 7873.349926
#> LO-REDUCTION    1163 7873.351113 7873.349926
#> LO-REDUCTION    1165 7873.351074 7873.349926
#> REFLECTION      1167 7873.350719 7873.349709
#> REFLECTION      1169 7873.350401 7873.349570
#> EXTENSION       1171 7873.350307 7873.349441
#> HI-REDUCTION    1173 7873.350209 7873.349441
#> EXTENSION       1175 7873.350163 7873.349170
#> LO-REDUCTION    1177 7873.350152 7873.349170
#> EXTENSION       1179 7873.349981 7873.348487
#> LO-REDUCTION    1181 7873.349926 7873.348487
#> LO-REDUCTION    1183 7873.349841 7873.348487
#> LO-REDUCTION    1185 7873.349709 7873.348487
#> LO-REDUCTION    1187 7873.349570 7873.348487
#> EXTENSION       1189 7873.349441 7873.348184
#> EXTENSION       1191 7873.349188 7873.347291
#> LO-REDUCTION    1193 7873.349170 7873.347291
#> LO-REDUCTION    1195 7873.348920 7873.347291
#> LO-REDUCTION    1197 7873.348713 7873.347291
#> LO-REDUCTION    1199 7873.348667 7873.347291
#> REFLECTION      1201 7873.348497 7873.347168
#> LO-REDUCTION    1203 7873.348487 7873.347168
#> REFLECTION      1205 7873.348184 7873.347091
#> EXTENSION       1207 7873.347835 7873.346777
#> EXTENSION       1209 7873.347816 7873.346464
#> LO-REDUCTION    1211 7873.347670 7873.346464
#> LO-REDUCTION    1213 7873.347454 7873.346464
#> REFLECTION      1215 7873.347361 7873.346413
#> REFLECTION      1217 7873.347291 7873.346256
#> REFLECTION      1219 7873.347168 7873.346148
#> LO-REDUCTION    1221 7873.347091 7873.346148
#> HI-REDUCTION    1223 7873.346939 7873.346148
#> REFLECTION      1225 7873.346777 7873.345928
#> LO-REDUCTION    1227 7873.346508 7873.345928
#> LO-REDUCTION    1229 7873.346464 7873.345928
#> LO-REDUCTION    1231 7873.346458 7873.345928
#> LO-REDUCTION    1233 7873.346413 7873.345928
#> LO-REDUCTION    1235 7873.346256 7873.345928
#> LO-REDUCTION    1237 7873.346222 7873.345928
#> LO-REDUCTION    1239 7873.346167 7873.345928
#> EXTENSION       1241 7873.346148 7873.345718
#> LO-REDUCTION    1243 7873.346064 7873.345718
#> LO-REDUCTION    1245 7873.346049 7873.345718
#> LO-REDUCTION    1247 7873.346046 7873.345718
#> LO-REDUCTION    1249 7873.345997 7873.345718
#> LO-REDUCTION    1251 7873.345996 7873.345718
#> LO-REDUCTION    1253 7873.345940 7873.345718
#> LO-REDUCTION    1255 7873.345928 7873.345718
#> EXTENSION       1257 7873.345908 7873.345621
#> EXTENSION       1259 7873.345848 7873.345579
#> EXTENSION       1261 7873.345814 7873.345299
#> LO-REDUCTION    1263 7873.345750 7873.345299
#> LO-REDUCTION    1265 7873.345750 7873.345299
#> LO-REDUCTION    1267 7873.345741 7873.345299
#> LO-REDUCTION    1269 7873.345733 7873.345299
#> LO-REDUCTION    1271 7873.345718 7873.345299
#> EXTENSION       1273 7873.345621 7873.345176
#> EXTENSION       1275 7873.345579 7873.345023
#> LO-REDUCTION    1277 7873.345533 7873.345023
#> LO-REDUCTION    1279 7873.345496 7873.345023
#> EXTENSION       1281 7873.345478 7873.345003
#> REFLECTION      1283 7873.345427 7873.344920
#> REFLECTION      1285 7873.345367 7873.344868
#> REFLECTION      1287 7873.345299 7873.344845
#> LO-REDUCTION    1289 7873.345176 7873.344845
#> LO-REDUCTION    1291 7873.345165 7873.344845
#> REFLECTION      1293 7873.345057 7873.344721
#> HI-REDUCTION    1295 7873.345023 7873.344721
#> LO-REDUCTION    1297 7873.345003 7873.344721
#> LO-REDUCTION    1299 7873.344920 7873.344721
#> EXTENSION       1301 7873.344880 7873.344649
#> LO-REDUCTION    1303 7873.344874 7873.344649
#> LO-REDUCTION    1305 7873.344868 7873.344649
#> LO-REDUCTION    1307 7873.344862 7873.344649
#> LO-REDUCTION    1309 7873.344845 7873.344649
#> EXTENSION       1311 7873.344807 7873.344604
#> LO-REDUCTION    1313 7873.344775 7873.344604
#> LO-REDUCTION    1315 7873.344769 7873.344604
#> LO-REDUCTION    1317 7873.344760 7873.344604
#> EXTENSION       1319 7873.344722 7873.344489
#> LO-REDUCTION    1321 7873.344721 7873.344489
#> LO-REDUCTION    1323 7873.344677 7873.344489
#> LO-REDUCTION    1325 7873.344663 7873.344489
#> EXTENSION       1327 7873.344649 7873.344444
#> LO-REDUCTION    1329 7873.344648 7873.344444
#> EXTENSION       1331 7873.344628 7873.344373
#> LO-REDUCTION    1333 7873.344604 7873.344373
#> LO-REDUCTION    1335 7873.344553 7873.344373
#> LO-REDUCTION    1337 7873.344526 7873.344373
#> EXTENSION       1339 7873.344514 7873.344349
#> LO-REDUCTION    1341 7873.344512 7873.344349
#> REFLECTION      1343 7873.344489 7873.344300
#> LO-REDUCTION    1345 7873.344444 7873.344300
#> LO-REDUCTION    1347 7873.344431 7873.344300
#> EXTENSION       1349 7873.344403 7873.344240
#> LO-REDUCTION    1351 7873.344381 7873.344240
#> REFLECTION      1353 7873.344373 7873.344229
#> HI-REDUCTION    1355 7873.344362 7873.344229
#> EXTENSION       1357 7873.344349 7873.344169
#> LO-REDUCTION    1359 7873.344338 7873.344169
#> EXTENSION       1361 7873.344300 7873.344093
#> LO-REDUCTION    1363 7873.344300 7873.344093
#> LO-REDUCTION    1365 7873.344287 7873.344093
#> LO-REDUCTION    1367 7873.344261 7873.344093
#> LO-REDUCTION    1369 7873.344240 7873.344093
#> LO-REDUCTION    1371 7873.344229 7873.344093
#> Exiting from Nelder Mead minimizer
#>     1373 function evaluations used
#> Post processing for method  Nelder-Mead 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  Nelder-Mead 
#> Compute gradient approximation at finish of  Nelder-Mead 
#> Save results from method  Nelder-Mead 
#> $par
#> (Intercept)           P         pi1         pi2      theta5      theta6 
#>  2.98631522 -0.97538823  0.01743891  2.21266368  1.06191570  1.01343231 
#>      theta7      theta8 
#>  0.26143557  1.09163689 
#> 
#> $value
#> [1] 7873.344
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>     1373 
#> 
#> $gevals
#> gradient 
#>       NA 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] FALSE
#> 
#> $kkt2
#> [1] TRUE
#> 
#> $xtimes
#> user.self 
#>     1.062 
#> 
#> Assemble the answers

# use method L-BFGS-B instead of Nelder-Mead and print report every 50 iterations
l4 <- latentIV(y ~ P, optimx.args = list(method = "L-BFGS-B", control=list(trace = 2, REPORT=50)),
               data = dataLatentIV)

#> No start parameters were given. The linear model y ~ P is fitted to derive them.

#> The start parameters c((Intercept)=2.915, P=-0.852, pi1=0.589, pi2=2.001, theta5=0.5, theta6=1, theta7=0.5, theta8=1) are used for optimization.

#> fn is  fn 
#> Looking for method =  L-BFGS-B 
#> Methods to be used:[1] "L-BFGS-B"
#> Function has  8  arguments
#> Analytic gradient not made available.
#> Analytic Hessian not made available.
#> Scale check -- log parameter ratio= 0.7657037   log bounds ratio= NA 
#> optcfg:$fname
#> [1] "fn"
#> 
#> $npar
#> [1] 8
#> 
#> $ctrl
#> $ctrl$follow.on
#> [1] FALSE
#> 
#> $ctrl$save.failures
#> [1] TRUE
#> 
#> $ctrl$trace
#> [1] 2
#> 
#> $ctrl$kkt
#> [1] TRUE
#> 
#> $ctrl$all.methods
#> [1] FALSE
#> 
#> $ctrl$starttests
#> [1] TRUE
#> 
#> $ctrl$maximize
#> [1] FALSE
#> 
#> $ctrl$dowarn
#> [1] TRUE
#> 
#> $ctrl$usenumDeriv
#> [1] FALSE
#> 
#> $ctrl$kkttol
#> [1] 0.001
#> 
#> $ctrl$kkt2tol
#> [1] 1e-06
#> 
#> $ctrl$badval
#> [1] 8.988466e+307
#> 
#> $ctrl$scaletol
#> [1] 3
#> 
#> $ctrl$REPORT
#> [1] 50
#> 
#> $ctrl$have.bounds
#> [1] FALSE
#> 
#> 
#> $usenumDeriv
#> [1] FALSE
#> 
#> $ufn
#> function (params, m.data.mvnorm, use.intercept, name.intercept, 
#>     name.endo.param) 
#> {
#>     if (use.intercept) {
#>         b00 <- params[name.intercept]
#>         a1 <- params[name.endo.param]
#>     }
#>     else {
#>         b00 <- 0
#>         a1 <- params[name.endo.param]
#>     }
#>     pt <- exp(params["theta5"])
#>     pt <- pt/(1 + pt)
#>     m.sigma <- matrix(c(params["theta6"], params["theta7"], params["theta7"], 
#>         params["theta8"]), byrow = TRUE, ncol = 2, nrow = 2)
#>     s2e <- m.sigma[1, 1]
#>     s2v <- m.sigma[2, 2]
#>     sev <- m.sigma[1, 2]
#>     varcov <- matrix(0, 2, 2)
#>     varcov[1, 1] <- a1 * a1 * s2v + 2 * a1 * sev + s2e
#>     varcov[2, 1] <- a1 * s2v + sev
#>     varcov[1, 2] <- varcov[2, 1]
#>     varcov[2, 2] <- s2v
#>     pi1 <- params["pi1"]
#>     mu1 <- matrix(data = c(b00 + a1 * pi1, pi1), nrow = 2, ncol = 1)
#>     pi2 <- params["pi2"]
#>     mu2 <- matrix(c(b00 + a1 * pi2, pi2), nrow = 2, ncol = 1)
#>     log.pdf1 <- dmvnorm(m.data.mvnorm, mean = mu1, sigma = varcov, 
#>         log = TRUE)
#>     log.pdf2 <- dmvnorm(m.data.mvnorm, mean = mu2, sigma = varcov, 
#>         log = TRUE)
#>     max.AB <- pmax(log(pt) + log.pdf1, log(1 - pt) + log.pdf2)
#>     logLL.lse <- sum(max.AB + log(pt * exp(log.pdf1 - max.AB) + 
#>         (1 - pt) * exp(log.pdf2 - max.AB)))
#>     return(-1 * logLL.lse)
#> }
#> <bytecode: 0x7fc0fa718390>
#> <environment: namespace:REndo>
#> 
#> $have.bounds
#> [1] FALSE
#> 
#> $method
#> [1] "L-BFGS-B"
#> 
#> Method:  L-BFGS-B 
#> N = 8, M = 5 machine precision = 2.22045e-16
#> This problem is unconstrained.
#> 
#> iterations 28
#> function evaluations 32
#> segments explored during Cauchy searches 1
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 0
#> norm of the final projected gradient 0.0630359
#> final function value 7872.1
#> 
#> final  value 7872.102327 
#> converged
#> Post processing for method  L-BFGS-B 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  L-BFGS-B 
#> Compute gradient approximation at finish of  L-BFGS-B 
#> Save results from method  L-BFGS-B 
#> $par
#> (Intercept)           P         pi1         pi2      theta5      theta6 
#>  2.98907680 -0.97725302 -0.02486042  2.13742890  0.92538516  1.00268565 
#>      theta7      theta8 
#>  0.24989801  1.04266832 
#> 
#> $value
#> [1] 7872.102
#> 
#> $message
#> [1] "CONVERGENCE: REL_REDUCTION_OF_F <= FACTR*EPSMCH"
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>       32 
#> 
#> $gevals
#> gradient 
#>       32 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] TRUE
#> 
#> $xtimes
#> user.self 
#>     0.412 
#> 
#> Assemble the answers

# read out all coefficients, incl auxiliary coefs
lat.all.coefs <- coef(l4)
# same as above
lat.all.coefs <- coef(l4, complete = TRUE)
# only main model coefs
lat.main.coefs <- coef(l4, complete = FALSE)
# }